### 2024-04-05 #6582 @editorialbot
[PRE REVIEW]: MITgcm.jl: a Julia Interface to the MITgcm
**Submitting author:** @gaelforget (<a href="http://orcid.org/0000-0002-4234-056X">Gael Forget</a>)
**Repository:** https://github.com/gaelforget/MITgcm.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.2
