### 2024-04-05 #6582 @kthyng
Comments/questions:
* are there installation instructions somewhere?
* your paper needs space between references in a list
