### 2024-04-06 ea45b1c
add installation in docs
