+++ docs/make.jl
@@ -9,8 +9,9 @@ makedocs(;
         "Home" => "index.md",
         "Manual" => "functionalities.md",
         "Examples" => "examples.md",
-        "API Reference" => "functionalities_more.md",
-        "Contributing" => "contributing.md",
+        "Reference" => "functionalities_more.md",
+        "Contribute" => "contributing.md",
+        "Install" => "install.md",
     ],
     warnonly = [:cross_references,:missing_docs],
     repo="https://github.com/gaelforget/MITgcm.jl/blob/{commit}{path}#L{line}",
