+++ docs/joss/paper.bib
@@ -14,7 +14,7 @@ year = {2015}
 }
 
 @software{Cbiomes2019,
-  author       = {Gael Forget},
+  author       = {Ga{\"{e}}l Forget},
   title        = {CBIOMES-global: a global ocean state estimate for marine ecosystems.},
   month        = apr,
   year         = 2019,
@@ -25,7 +25,7 @@ year = {2015}
 }
 
 @software{ClimateModels2024,
-  author       = {Gael Forget},
+  author       = {Ga{\"{e}}l Forget},
   title        = {ClimateModels.jl: Julia interface to climate models, and tracked modeling workflow framework.},
   month        = apr,
   year         = 2024,
@@ -43,12 +43,12 @@ year = {2015}
   volume = {6},
   number = {60},
   pages = {2813},
-  author = {Gael Forget},
+  author = {Ga{\"{e}}l Forget},
   title = {IndividualDisplacements.jl: a Julia package to simulate and study particle displacements within the climate system},
   journal = {Journal of Open Source Software}
 }
 
-@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Gael Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
+@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Ga{\"{e}}l Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
 
 @article{Duckworth2023,
 	author = {Barbara Duckworth and Stephanie Dutkiewicz and Oliver Jahn and Christopher N. Hill and Christopher L. Follett},
@@ -75,7 +75,7 @@ year = {2015}
 }
 
 @article{Rousselet2021,
-author = {Louise Rousselet  and Paola Cessi  and Gael Forget },
+author = {Louise Rousselet  and Paola Cessi  and Ga{\"{e}}l Forget },
 title = {Coupling of the mid-depth and abyssal components of the global overturning circulation according to a state estimate},
 journal = {Science Advances},
 volume = {7},
@@ -110,7 +110,7 @@ year = {2019}
   journal = {Journal of Open Source Software}
 }
 
-@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Gael Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
+@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Ga{\"{e}}l Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
 
 @article{Forget2021,
   doi = {10.21105/joss.02813},
@@ -120,13 +120,13 @@ year = {2019}
   volume = {6},
   number = {60},
   pages = {2813},
-  author = {Gael Forget},
[…]
