+++ docs/joss/paper.md
@@ -20,17 +20,17 @@ General circulation models are used to study climate, ocean and atmosphere dynam
 
 # Statement of Need
 
-The cutting-edge of climate modeling, and much of its legacy, is based on numerical models written in compiled languages like `Fortran` or `C`. The MIT general circulation model (`MITgcm`) for example has recently been run globally at the kilometer scale to provide an unprecented view of ocean dynamics (Fig.~\ref{fig:examples}, @camp-etal:04, @marshall1997fvi). With its unrivaled adjoint modeling capabilities, `MITgcm` is also the computational engine that powers the ECCO reanalysis, a data-assimilating product widely-used in climate science (Fig. \ref{fig:examples}, @Forget2015a, @Forget2024). `MITgcm` further provides unique modeling capabilities for ocean biogeochemistry, ecology, and optics (@Dutkiewicz2015,@Cbiomes2019). While a new generation of models, written in languages like C++ and Julia, is poised to better exploit GPUs (@OceananigansJOSS,@Wu2022,@e3sm-model), `Fortran`-based models are expected to remain popular on other computer architectures for the foreseable future. They also provide a crucial reference point to evaluate next generation models.
+The cutting-edge of climate modeling, and much of its legacy, is based on numerical models written in compiled languages like `Fortran` or `C`. The MIT general circulation model (`MITgcm`) for example has recently been run globally at the kilometer scale to provide an unprecented view of ocean dynamics (Fig.~\ref{fig:examples}, @camp-etal:04,  @marshall1997fvi). With its unrivaled adjoint modeling capabilities, `MITgcm` is also the computational engine that powers the ECCO reanalysis, a data-assimilating product widely-used in climate science (Fig. \ref{fig:examples},  @Forget2015a,  @Forget2024). `MITgcm` further provides unique modeling capabilities for ocean biogeochemistry, ecology, and optics (@Dutkiewicz2015, @Cbiomes2019). While a new generation of models, written in languages like C++ and Julia, is poised to better exploit GPUs (@OceananigansJOSS, @Wu2022, @e3sm-model), `Fortran`-based models are expected to remain popular on other computer architectures for the foreseable future. They also provide a crucial reference point to evaluate next generation models.
 
 Models like `MITgcm` benefit from decades of accumulated expertise, careful inspection, and continuous integration. Running such a model or exploiting its results does not require knowing the computer language being used internally. However, `Fortran`-based models can sometime appear complicated or inconvenient to operate due to technical hurdles like having to use a compiler directly, to edit text files manually, or to deal with shell-scripting. Fortunately, such issues are easily alleviated by providing a user-friendly interface (e.g., Fig. \ref{fig:interact}) written in a high-level language (e.g., `Julia`) to interact with climate models written in lower-level languages as done here for `MITgcm`.
 
[…]
