### 2024-04-06 #6582 @gaelforget
> Comments/questions:
>
> * are there installation instructions somewhere?

added in https://github.com/gaelforget/MITgcm.jl/commit/ea45b1c26635873ffb10118f6dca1eb30827d9f6

> * your paper needs space between references in a list

does this work ?

https://github.com/gaelforget/MITgcm.jl/commit/6b15663aee7bae6583f848a366f2a87f2726b1cf
