+++ v0.3.3/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"functionalities/#Manual","page":"Manual","title":"Manual","text":"","category":"section"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"MITgcm.jl provides a suite of tools for analyzing MITgcm results, compiling the model, modifying its inputs, running simulations, etc all from within julia. Much of the functionalities are documented via the Examples section, as highlighted hereafter.","category":"page"},{"location":"functionalities/#Read-/-Write-MITgcm-Files","page":"Manual","title":"Read / Write MITgcm Files","text":"","category":"section"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"Let's assume that user has run an MITgcm configuration in standard fashion, and wants to examine the output in Julia. This is one of the common use case for MITgcm.jl.","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"The tutorial_global_oce_biogeo, which comes with MITgcm, provides a representative example. With MITgcmTool.jl you can just run it in a notebook (MITgcm_tutorial_global_oce_biogeo.jl).","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"note: Note\nAlternatively, the MITgcm documentation explains how to complile and run tutorials at the linux command line. ","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"Once the model has run, it's output is found in the run/ folder (see notebook). The scan_rundir function can provide a summary information about files in the run/ folder.","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"rundir=joinpath(pth,\"run\")\nsc=scan_rundir(rundir)\nsc.completed","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"The result will indicate whether the model run has complete normally based upon the Standard Output file (output.txt or STDOUT.000). ","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"The scan_rundir function checks whether the model output is format binary (sc.params_files.use_mdsio) or NetCDF (sc.params_files.use_mnc). It also identifies the type of grid that was used (sc.params_grid).","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"With this information, we are ready to read model output. The various files that MITgcm can generate are covered below.","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"Scan Run Folder\nMDS Files (binary)\nMNC Files (netcdf)\nGrid Files (binary or netcdf)\nStandard Output (text)\nInput Files (text)\nOther Files","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"Grid variables are often needed for analysis. The grid output can be read from file using either GridLoad_mdsio or GridLoad_mnc. This will return Γ.XC, Γ.YC, etc formated using MeshArrays.jl. See also GridLoad_native.","category":"page"},{"location":"functionalities/","page":"Manual","title":"Manual","text":"note:
[…]
