### 2024-04-12 #6582 @gaelforget
> @gaelforget That would be great. Thanks!

Here is a list of potential reviewers by group :

MITgcm community active on github : jm-c, christophernhill, mjlosch, owang01 , jrscott, jahn, averdy, ifenty, menemenlis , hongandyan , emmomp , IvanaEscobar

Julia community in related fields : simone-silvestri , ali-ramadhan, briochemc , Alexander-Barth , visr  , barbara42 , glwagner, navidcy , jbisits , swilliamson7 , fonsp , SimonDanisch , Datseris , felixcremer

From [JOSS Reviewers site](https://reviewers.joss.theoj.org/lookup) : JordiBolibar , patrickcgray ,

ps. I know many of the listed people to various degrees, just fyi. They might be able to suggest others.
