+++ docs/make.jl
@@ -7,9 +7,10 @@ makedocs(;
     format=Documenter.HTML(),
     pages=[
         "Home" => "index.md",
-        "Manual" => "functionalities.md",
+        "Manual" => ["functionalities.md","functionalities_interface.md",
+        "functionalities_configurations.md","functionalities_read.md",
+        "functionalities_more.md"],
         "Examples" => "examples.md",
-        "Reference" => "functionalities_more.md",
         "Contribute" => "contributing.md",
         "Install" => "install.md",
     ],
@@ -21,7 +22,7 @@ makedocs(;
 
 pth = joinpath(@__DIR__, "build","examples")
 lst=("HS94_animation.jl","HS94_particles.jl","MITgcm_configurations.jl","MITgcm_run.jl","MITgcm_worklow.jl","HS94_Makie.jl","MITgcm_scan_output.jl")
-for i in lst
+if false #for i in lst
     fil_in=joinpath(@__DIR__,"..","examples",i)
     fil_out=joinpath(pth,i[1:end-2]*"html")
     PlutoSliderServer.export_notebook(fil_in)
