+++ docs/src/functionalities_more.md
@@ -1,10 +1,6 @@
+# More
 
-### API Reference
-
-```@index
-```
-
-### Format Conversions
+## Format Conversions
 
 ```@docs
 findtiles
@@ -14,7 +10,7 @@ convert2array
 convert2gcmfaces
 ```
 
-### Formulae etc
+## Formulae etc
 
 ```@docs
 SeaWaterDensity
