+++ docs/make.jl
@@ -22,7 +22,7 @@ makedocs(;
 
 pth = joinpath(@__DIR__, "build","examples")
 lst=("HS94_animation.jl","HS94_particles.jl","MITgcm_configurations.jl","MITgcm_run.jl","MITgcm_worklow.jl","HS94_Makie.jl","MITgcm_scan_output.jl")
-if false #for i in lst
+for i in lst
     fil_in=joinpath(@__DIR__,"..","examples",i)
     fil_out=joinpath(pth,i[1:end-2]*"html")
     PlutoSliderServer.export_notebook(fil_in)
