+++ docs/src/functionalities.md
@@ -1,4 +1,15 @@
-# Main Functionalities
+# Overview
+
+## Installation Instructions
+
+You can install the latest version of `MITgcm.jl` using the built-in [package manager](https://pkgdocs.julialang.org/). 
+
+```
+using Pkg
+Pkg.add("MITgcm")
+```
+
+## Main Functionalities
 
 `MITgcm.jl` provides a suite of tools for analyzing [MITgcm](https://mitgcm.readthedocs.io/en/latest/?badge=latest) results, compiling the model, modifying its inputs, running simulations, and more -- from within `julia`. Functionalities are also documented via the [Examples](@ref) section.
 
@@ -7,7 +18,7 @@
 ```@index
 ```
 
-## Read / Write MITgcm Files
+## MITgcm File Formats
 
 A common use case for `MITgcm.jl` is : we have output from a previous `MITgcm` run available, and want to analyze it in `Julia`. 
 
@@ -43,7 +54,7 @@ Grid variables are often needed for analysis. The grid output can be read from f
 !!! note 
     [MITgcm\_scan\_output.jl](https://github.com/gaelforget/MITgcm.jl/blob/master/examples/MITgcm_scan_output.jl) does this in bulk for all configurations in `MITgcm/verification` and further displays each grid ([this page](https://gaelforget.github.io/MITgcm.jl/dev/examples/MITgcm_scan_output.html)).
 
-## MITgcm Configurations
+## Model Configurations
 
 The [`verification_experiments`](@ref) function lists standard model configurations found in the `joinpath(MITgcm_path[1],"verification")` folder. 
 
