+++ docs/src/install.md
@@ -1,8 +1,0 @@
-# Installation instructions
-
-You can install the latest version of `MITgcm.jl` using the built-in [package manager](https://pkgdocs.julialang.org/). 
-
-```
-using Pkg
-Pkg.add("MITgcm")
-```
