+++ docs/src/examples.md
@@ -2,12 +2,10 @@
 using MITgcm
 ```
 
-# Examples
-
 !!! note
 	Compiling MITgcm requires [a fortran compiler](https://fortran-lang.org/learn/os_setup/install_gfortran). This is a requirement for all notebooks except `MITgcm_configurations.jl`.
 
-## Standard Workflow
+# Standard Workflow
 
 ```@example 1
 using MITgcm
@@ -26,7 +24,7 @@ MITgcm_launch(MC)
 readdir(MC,"run")
 ```
 
-## Notebook Examples
+# Notebook Examples
 
 Below are links to static html versions of the examples which one can open with a web browser.
 
@@ -52,7 +50,7 @@ include(notebook)
 '
 ```
 
-## Notebooks Listing
+# Notebooks Listing
 
 - [MITgcm_configurations.jl](MITgcm_configurations.html) ([code link](https://raw.githubusercontent.com/gaelforget/MITgcm.jl/master/examples/MITgcm_configurations.jl)); explore MITgcm configurations and their parameters.
 - [MITgcm\_scan\_output.jl](MITgcm_scan_output.html) ([code link](https://raw.githubusercontent.com/gaelforget/MITgcm.jl/master/examples/MITgcm_scan_output.jl)) : scan run directory, standard output, read grid files, and vizualize. 
