+++ docs/src/functionalities_more.md
@@ -1,4 +1,4 @@
-# More
+# More Functionalities
 
 ## Format Conversions
 
