+++ docs/src/functionalities_interface.md
@@ -1,13 +1,16 @@
 # ClimateModels Interface
 
-This allows us to create model simulations using `MITgcm` conveniently from `Julia`.
+The [ClimateModels.jl]()'s interface is implemented in this pacakge. This allows you to easily create model simulations using `MITgcm` conveniently from `Julia`.
 
-```
+```@example 1
+using MITgcm
 MC=MITgcm_config(configuration="advect_xy")
 setup(MC)
-build(MC)
+build(MC,"--allow-skip")
 MITgcm_launch(MC)
 ```
+
+## Functionalities
 
 ```@docs
 setup
