+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"functionalities/#Overview","page":"Overview","title":"Overview","text":"","category":"section"},{"location":"functionalities/#Installation-Instructions","page":"Overview","title":"Installation Instructions","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"You can install the latest version of MITgcm.jl using the built-in package manager. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"using Pkg\nPkg.add(\"MITgcm\")","category":"page"},{"location":"functionalities/#Main-Functionalities","page":"Overview","title":"Main Functionalities","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"MITgcm.jl provides a suite of tools for analyzing MITgcm results, compiling the model, modifying its inputs, running simulations, and more – from within julia. Functionalities are also documented via the Examples section.","category":"page"},{"location":"functionalities/#API-Reference","page":"Overview","title":"API Reference","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"","category":"page"},{"location":"functionalities/#MITgcm-File-Formats","page":"Overview","title":"MITgcm File Formats","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"A common use case for MITgcm.jl is : we have output from a previous MITgcm run available, and want to analyze it in Julia. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"The tutorial_global_oce_biogeo, which comes with MITgcm, provides a representative example. With MITgcmTool.jl you can just run it in a notebook (MITgcm_tutorial_global_oce_biogeo.jl).","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"note: Note\nAlternatively, the MITgcm documentation explains how to complile and run tutorials at the linux command line. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Once the model has run, it's output is found in the run/ folder (see notebook). The scan_rundir function can provide a summary information about files in the run/ folder.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"rundir=joinpath(pth,\"run\")\nsc=scan_rundir(rundir)\nsc.completed","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"The result will indicate whether the model run has complete normally based upon the Standard Output file (output.txt or STDOUT.000). ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"The scan_rundir function checks whether the model output is format binary (sc.params_files.use_mdsio) or NetCDF (sc.params_files.use_mnc). It also identifies the type of grid that was used (sc.params_grid).","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"With this information, we are ready to read model output. The various files that MITgcm can
[…]
