+++ v0.3.4/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"functionalities/#Overview","page":"Overview","title":"Overview","text":"","category":"section"},{"location":"functionalities/#Installation-Instructions","page":"Overview","title":"Installation Instructions","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"You can install the latest version of MITgcm.jl using the built-in package manager. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"using Pkg\nPkg.add(\"MITgcm\")","category":"page"},{"location":"functionalities/#Main-Functionalities","page":"Overview","title":"Main Functionalities","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"MITgcm.jl provides a suite of tools for analyzing MITgcm results, compiling the model, modifying its inputs, running simulations, and more – all from within julia. Functionalities are also documented via the Examples section.","category":"page"},{"location":"functionalities/#MITgcm-File-Formats","page":"Overview","title":"MITgcm File Formats","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"A common use case for MITgcm.jl is to use and analyze model output from a previous MITgcm run. As an example, the notebook from JuliaCon2021 (MITgcm_tutorial_global_oce_biogeo.jl) reads and visualize results from the standard MITgcm configuration called tutorial_global_oce_biogeo.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"note: Note\nThis notebook builds and runs tutorial_global_oce_biogeo from within Julia. Alternatively, the MITgcm documentation explains how to build and run tutorials at the command line in linux. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"scan_rundir / scan_stdout provides a summary information about what's in the model run/ folder. With this information, we are ready to read model output. The various files that MITgcm can generate are covered below.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Standard Output (text)\nInput Files (text)\nMDS Files (binary output)\nMNC Files (netcdf output)\nGrid Files (binary or netcdf)\nOther Files","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Grid variables are often needed for analysis. The grid output can be read from file using either GridLoad_mdsio or GridLoad_mnc. This will return Γ.XC, Γ.YC, etc formated using MeshArrays.jl. See also GridLoad_native.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"note: Note\nThe MITgcm_scan_output.jl notebook does this in bulk for all configurations in MITgcm/verification and displays the gridded model domain for each model configuration (this page).","category":"page"},{"location":"functionalities/#Model-Configurations","page":"Overview","title":"Model Configurations","text":"","category":"section"},{"location":"function
[…]
