### 2024-04-14 b24a102
start generalize, beyond verif config
