+++ src/MITgcm.jl
@@ -20,7 +20,7 @@ export testreport, build, compile, setup, clean
 #export pause, stop, clock, monitor, train, help
 export verification_experiments, read_namelist, write_namelist
 export read_mdsio, read_meta, read_available_diagnostics
-export scan_rundir, scan_stdout
+export read_toml, scan_rundir, scan_stdout
 export read_bin, read_flt, read_mnc, read_nctiles, findtiles
 export GridLoad_mnc, GridLoad_mdsio, GridLoad_native
 export cube2compact, compact2cube, convert2array, convert2gcmfaces
