### 2024-04-15 a1b2eb6
add configs, namelist treatment
