### 2024-04-15 f8ec52f
separate out write_all_namelists
