+++ examples/configurations/prepare_toml.jl
@@ -1,8 +1,8 @@
 
 using MITgcm, TOML, OrderedCollections
 
-params=MITgcm.read_namelist_files("input"); nam="ECCO4"
-#params=MITgcm.read_namelist_files("run"); nam="OCCA2"
+params=read_all_namelists("input"); nam="ECCO4"
+#params=read_all_namelists("run"); nam="OCCA2"
 
 P=OrderedDict()
 P[:main]=OrderedDict(
