+++ docs/src/functionalities_read.md
@@ -37,7 +37,8 @@ GridLoad_native
 MITgcm_namelist
 read_namelist
 write_namelist
-read_namelist_files
+read_all_namelists
+write_all_namelists
 read_toml
 ```
 
