### 2024-04-15 49ea5e0
generalize setup, improve docs on configs
