+++ src/Types.jl
@@ -37,10 +37,8 @@ which contains
 ```
     model :: String = "MITgcm"
     configuration :: String = ""
-    options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
-    status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     channel :: Channel{Any} = Channel{Any}(10) 
     folder :: String = tempdir()
     ID :: UUID = UUIDs.uuid4()
@@ -55,7 +53,6 @@ unknown_config=MITgcm_config(configuration="unknown")
 Base.@kwdef struct MITgcm_config <: AbstractModelConfig
     model :: String = "MITgcm"
     configuration :: String = ""
-    options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
     status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
