+++ src/MITgcm.jl
@@ -18,10 +18,11 @@ export MITgcm_download, HS94_pickup_download
 export MITgcm_config, MITgcm_namelist, MITgcm_launch
 export testreport, build, compile, setup, clean
 #export pause, stop, clock, monitor, train, help
-export verification_experiments, read_namelist, write_namelist
+export verification_experiments, verification_setup
+export read_namelist, write_namelist, read_toml
 export read_all_namelists, write_all_namelists
 export read_mdsio, read_meta, read_available_diagnostics
-export read_toml, scan_rundir, scan_stdout
+export scan_rundir, scan_stdout
 export read_bin, read_flt, read_mnc, read_nctiles, findtiles
 export GridLoad_mnc, GridLoad_mdsio, GridLoad_native
 export cube2compact, compact2cube, convert2array, convert2gcmfaces
