+++ docs/src/functionalities_configurations.md
@@ -1,5 +1,13 @@
 # Model Configurations
 
+## MITgcm_config
+
+```@docs
+MITgcm_config
+```
+
+## Verification Experiments
+
 Get the default version of `MITgcm` source code, and locate the local copy:
 
 ```@example 1
@@ -26,5 +34,5 @@ MITgcm_config(configuration="MLAdjust")
 MITgcm_download
 MITgcm_path
 verification_experiments
-MITgcm_config
+verification_setup
 ```
