### 2024-04-15 8613348
generalize build, add ECCO4 as example
