+++ src/ModelSteps.jl
@@ -59,7 +59,7 @@ function testreport(config::MITgcm_config,ext="")
     return true
 end
 
-import ClimateModels: compile, build, setup, clean
+import ClimateModels: compile, build, setup, clean, git
 
 """
     clean(config::MITgcm_config)
@@ -89,20 +89,13 @@ code files, headers, etc  in the `build/` folder before compiling the model
 (part of the climate model interface as specialized for `MITgcm`)
 """
 function build(config::MITgcm_config)
-    nam=config.configuration
     try
         pth=pwd()
     catch e
         cd()
     end
     pth=pwd()
-
-    if !haskey(config.inputs,:setup)||(config.inputs[:setup][:main][:category]=="verification")
-        cd("$(MITgcm_path[1])/verification/$(nam)/build")
-    else
-        error("unknown configuration category")
-    end
-
+    cd(config.inputs[:setup][:build][:path])
     try
         @suppress run(`../../../tools/genmake2 -mods=../code`) #$ext
         @suppress run(`make clean`)
@@ -125,9 +118,8 @@ code files, headers, etc  in the `build/` folder before `make` compiles the mode
 (part of the climate model interface as specialized for `MITgcm`)
 """
 function build(config::MITgcm_config,options::String)
-    nam=config.configuration
     if options=="--allow-skip"
-        tst=!isfile(joinpath(MITgcm_path[1],"verification",nam,"build","mitgcmuv"))
+        tst=!isfile(joinpath(config.inputs[:setup][:build][:path],"mitgcmuv"))
         tst ? build(config) : nothing
     else
         build(config)
@@ -143,14 +135,13 @@ Compile the model using `make` in `build/` that has already been setup.
 (part of the climate model interface as specialized for `MITgcm`)
 """
 function compile(config::MITgcm_config)
-    nam=config.configuration
     try
         pth=pwd()
     catch e
         cd()
     end
     pth=pwd()
-    cd("$(MITgcm_path[1])/verification/$(nam)/build")
+    cd(config.inputs[:setup][:build][:path])
     try
         @suppress run(`make -j 4`)
     catch e
@@ -177,13 +168,11 @@ function setup(config::MITgcm_config)
     pth_run=joinpath(config.folder,string(config.ID),"run")
     !isdir(pth_run) ? mkdir(pth_run) : nothing
 
-    pth_log=joinpath(config.folder,string(config.ID),"log","tracked_parameters")
-    pth_mv=joinpath(config.folder,string(config.ID),"original_parameters")
-
+    pth_log=joinpath(config.folder,string(config.ID),"log")
     !isdir(pth_log) ? ClimateModels.git_log_init(config) : nothing
 
-    if !isfile(joinpath(pth_run,"data"))&&isfile(joinpath(pth_log,"data"))
-        p=pth_log
+    if !isfile(joinpath(pth_run,"data"))&&isfile(joinpath(pth_log,"tracked_parameters","data"))
+        p=joinpath(pth_log,"tracked_parameters")
         f=readdir(p)
         [symlink(joinpath(p,f[i]),joinpath(pth_run,f[i])) for i in 1:length(f)]
     end
@@ -192,21 +181,28 @@ function setup(config::MITgcm_config)
         params=verification_setup(config)
         push!(config.inputs,params...)
     elseif !isempty(config.inputs)
+        nam=config.inputs[:setup][:main][:name]
+        if nam=="ECCO4"||nam=="OCCA2"
+            setup_ECCO4!(config)
+        else
+            error("unknown model configuration")
+        end
[…]
