### 2024-04-15 #6582 @AnjaliSandip
👋 @mjlosch, @christophernhill, @ali-ramadhan & @simone-silvestri, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
