### 2024-04-15 #6582 @christophernhill
Happy to - I’ll take a look.

Chris

On Mon, Apr 15, 2024 at 16:49 Anjali Sandip ***@***.***>
wrote:

> 👋 @mjlosch <https://github.com/mjlosch>, @christophernhill
> <https://github.com/christophernhill>, @ali-ramadhan
> <https://github.com/ali-ramadhan> & @simone-silvestri
> <https://github.com/simone-silvestri>, would you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html.
>
