### 2024-04-20 186c0c9
document docker, binder solutions
