### 2024-04-20 646fe3d
rename as ModelConfigurations, mv ECCO4 code
