+++ src/MITgcm.jl
@@ -12,6 +12,7 @@ include("ModelSteps.jl")
 include("FormatConversions.jl")
 include("PhysicalOceanography.jl")
 include("ModelConfigurations.jl")
+include("ShellScripting.jl")
 
 export MITgcm_path, MITgcmScratchSpaces
 export MITgcm_download, HS94_pickup_download
