### 2024-04-21 7b00d25
mv testreport bits to src/ModelConfigurations.jl
