+++ Project.toml
@@ -6,23 +6,31 @@ version = "0.3.6"
 [deps]
 ClimateModels = "f6adb021-9183-4f40-84dc-8cea6f651bb0"
 CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
+Dataverse = "9c0b9be8-e31e-490f-90fe-77697562404d"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FortranFiles = "c58ffaec-ab22-586d-bfc5-781a99fd0b10"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Scratch = "6c6a2e73-6563-6170-7368-637461726353"
+SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tar = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [compat]
 ClimateModels = "0.3"
 CodecZlib = "0.7"
+Dataverse = "0.2"
+Downloads = "1"
 FortranFiles = "0.6"
 Glob = "1"
 MeshArrays = "0.3"
 Scratch = "1"
+SparseArrays = "1"
+Statistics = "1"
 Tar = "1"
 julia = "^1.9"
