+++ src/MITgcm.jl
@@ -21,7 +21,7 @@ include("ShellScripting.jl")
 export MITgcm_path, MITgcmScratchSpaces
 export MITgcm_download, HS94_pickup_download
 export MITgcm_config, MITgcm_namelist, MITgcm_launch
-export testreport, build, compile, setup, clean
+export testreport, build, compile, setup, clean, launch
 #export pause, stop, clock, monitor, train, help
 export verification_experiments, setup_verification!, testreport
 export setup_ECCO4!, ECCO4_inputs, ECCO4_testreport
