### 2024-05-01 #6710 @editorialbot
[REVIEW]: MITgcm.jl: a Julia Interface to the MITgcm
**Submitting author:** @gaelforget (<a href="http://orcid.org/0000-0002-4234-056X">Gael Forget</a>)
**Repository:** https://github.com/gaelforget/MITgcm.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.0

@simone-silvestri and @swilliamson7, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
