### 2024-05-05 c562b04
move NetCDF functionalities to extension
