+++ ext/MITgcmNetCDFExt.jl
@@ -1,0 +1,170 @@
+module MITgcmNetCDFExt
+
+    using MITgcm, NetCDF
+    using MITgcm.Printf, MITgcm.MeshArrays, MITgcm.ClimateModels, MITgcm.Glob
+    import MITgcm: read_nctiles, read_mnc
+
+"""
+    read_nctiles(fileName,fldName,mygrid; I, eccoVersion4Release4=false, verbose=false)
+
+Read model output from NCTiles file and convert to MeshArray instance. Setting the keyword
+argument `eccoVersion4Release4=true` allows `read_nctiles` to read in ECCOv4r4 data which
+has a different file naming convention to previous versions.
+```
+mygrid=GridSpec("LatLonCap")
+fileName="nctiles_grid/GRID"
+Depth=read_nctiles(fileName,"Depth",mygrid)
+hFacC=read_nctiles(fileName,"hFacC",mygrid)
+hFacC=read_nctiles(fileName,"hFacC",mygrid,I=(:,:,1))
+```
+"""
+function read_nctiles(fileName::String,fldName::String,mygrid::gcmgrid;
+    I::Union{Missing,Tuple{Colon,Colon,Vararg{Union{Colon,Integer}}}}=missing,
+    eccoVersion4Release4=false,verbose=false)
+
+    if (mygrid.class!="LatLonCap")||(mygrid.ioSize!=[90 1170])
+        error("non-llc90 cases have not yet been tested with read_nctiles")
+    end
+
+    pth0=dirname(fileName)
+    nam=split(fileName,"/")[end]
+    isempty(nam) ? nam=split(fileName,"/")[end-1] : nothing
+    occursin(".nctiles",nam) ? nam=nam[1:end-8] : nothing
+
+    isdir(fileName) ? pth1=fileName : pth1=pth0
+    lst=readdir(pth1)
+    lst=lst[findall(occursin.(nam,lst).*occursin.(".nc",lst))]
+
+    fileIn=joinpath(pth1,lst[1])
+    fileRoot= eccoVersion4Release4 ? fileIn[1:end-11] : fileIn[1:end-8]
+    ntile=ClimateModels.ncgetatt(fileIn,"Global","ntile")
+
+    x = ClimateModels.ncread(fileIn,fldName)
+    s = [size(x,i) for i in 1:ndims(x)]
+    n=length(size(x))
+    start=ones(Int,n)
+    count=-ones(Int,n)
+
+    ~ismissing(I) && length(I)!=n ? error("ncdims v I inconsistency") : nothing
+    if ~ismissing(I)
+        k=findall([!isa(I[i],Colon) for i=1:length(I)])
+        j=[I[i] for i in k]
+        start[k]=j
+        count[k].=1
+        s[k].=1
+    end
+
+    nr=50 # number of depth levels, this should be accessed from data in case nr = 1
+    numFiles = length(glob("*.nc", pth0))
+    f=eccoVersion4Release4==false ? Array{Float64, n}[] : numFiles > 1 ?
+                                                          MeshArray(mygrid,mygrid.ioPrec,nr,numFiles) :
+                                                          MeshArray(mygrid,mygrid.ioPrec,nr)
+    if eccoVersion4Release4
+        fill!(f, NaN)
+        tiles=Tiles(mygrid,90,90)
+        year=pth0[findlast('/', pth0)+1:end]
+        months=vcat("0" .* string.(1:9), string.(10:12))
+        fileCounter = 0 # for indexing the time in the MeshArray
+        for month in months
+            fileIn = @sprintf("%s_%s_%s.nc", fileRoot, year, month)
+            if isfile(fileIn) #skip if no file
+                fileCounter += 1
[…]
