+++ Project.toml
@@ -21,6 +21,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tar = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
+[weakdeps]
+NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
+
+[extensions]
+MITgcmNetCDFExt = ["NetCDF"]
+
 [compat]
 ClimateModels = "0.3"
 CodecZlib = "0.7"
@@ -29,6 +35,7 @@ Downloads = "1"
 FortranFiles = "0.6"
 Glob = "1"
 MeshArrays = "0.3"
+NetCDF = "0.11"
 Scratch = "1"
 SparseArrays = "1"
 Statistics = "1"
