+++ test/Project.toml
@@ -1,7 +1,6 @@
 [deps]
-ClimateModels = "f6adb021-9183-4f40-84dc-8cea6f651bb0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
+NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
 OceanStateEstimation = "891f6deb-a4f5-4bc5-a2e3-1e8f649cdd2c"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
