### 2024-05-06 1a087ec
remove old plot recipes
