### 2024-05-13 2bd95e8
use custom command if specified, mitgcmuv directly if otherwise
