### 2024-05-14 5bb8a56
improve ECCO example, launch command customization, create and use job.csh
