### 2024-05-14 c4451ee
streamline handling of parameters
