+++ docs/src/functionalities_read.md
@@ -1,8 +1,10 @@
 # MITgcm File Formats
 
-The two main output formats of MITgcm gridded fields are called `MDS` and `MNC`. In addition, the standard output is a text file that records events during the model run, including its successful completion.
+The two main output formats of MITgcm gridded fields are called `MDS` and `MNC`. In addition, the `standard output file` is a text file that records events during the model run, including its successful completion.
 
 ## Standard Output
+
+The `standard output file` can be scanned to collect information about the model run. 
 
 ```@docs
 scan_rundir
@@ -31,21 +33,23 @@ GridLoad_mnc
 GridLoad_native
 ```
 
-## Input Files
-
-```@docs
-MITgcm_namelist
-read_namelist
-write_namelist
-read_all_namelists
-write_all_namelists
-read_toml
-```
-
-## Other Files
+## Other Output Files
 
 ```@docs
 read_flt
 read_bin
 read_nctiles
+```
+
+# Input Files
+
+Run-time parameters to `MITgcm` are provided via text files. This package can read and write them in two formats (standard `TOML` format, or the native `MITgcm_namelist` format).
+
+```@docs
+read_toml
+MITgcm_namelist
+read_namelist
+write_namelist
+read_all_namelists
+write_all_namelists
 ```
