+++ docs/src/functionalities_configurations.md
@@ -2,6 +2,12 @@
 
 In `MITgcm.jl`, a model configuration is represented as a [`MITgcm_config`](@ref). Model parameters are handled as ordered dictionaries and stored as `TOML` files. Standard model configurations are readily supported.
 
+!!! note
+	Compiling MITgcm requires [a fortran compiler](https://fortran-lang.org/learn/os_setup/install_gfortran). Some configurations further require installing [MPI](https://mitgcm.readthedocs.io/en/latest/getting_started/getting_started.html?highlight=mpi_INC_DIR#building-with-mpi) and [NetCDF](https://mitgcm.readthedocs.io/en/latest/outp_pkgs/outp_pkgs.html?highlight=NetCDF#netcdf-i-o-pkg-mnc) libraries.
+	
+!!! tip
+	The [ECCO-Docker](https://github.com/gaelforget/ECCO-Docker#readme) _image_ has `MITgcm.jl` pre-installed, as well as `gfortran`, `MPI`, and `NetCDF` allowing to run any `MITgcm` configuration. The [ECCO-Binder](https://mybinder.org/v2/gh/gaelforget/ECCO-Docker/HEAD) _instance_ is available to try functionalities in the cloud, freely.
+
 ## MITgcm_config
 
 The data structure that enables `MITgcm.jl` is called `MITgcm_config`.
@@ -40,9 +46,9 @@ ECCO4_testreport.compare
 
 ## Verification Experiments
 
-The [MITgcm/verification](https://mitgcm.readthedocs.io/en/latest/getting_started/getting_started.html) folder is found in the `MITgcm` source code. It provides a suite of basic configurations used by model developers. 
+The [MITgcm/verification](https://mitgcm.readthedocs.io/en/latest/getting_started/getting_started.html) sub-folder of the `MITgcm` source code provides a suite of small model configurations, often used by model developers. 
 
-To use them, you want to download the `MITgcm` source code. For example :
+To use them, you first want to use [`MITgcm_download`](@ref) to get the source code.
 
 ```@example 1
 using MITgcm
@@ -50,14 +56,14 @@ MITgcm_download()
 MITgcm_path[1]
 ```
 
-To get the list of standard model configurations in `MITgcm/verification/` : 
+To list of these model configurations is provided by [`verification_experiments`](@ref). 
 
 ```@example 1
 ves=verification_experiments()
 [ve.configuration for ve in ves]
 ```
 
-To choose one : 
+To try one, you want to use [`MITgcm_config`](@ref) as follows.
 
 ```@example 1
 MITgcm_config(configuration="MLAdjust")
@@ -66,9 +72,9 @@ MITgcm_config(configuration="MLAdjust")
 ## Functionalities
 
 ```@docs
-setup_verification!
-verification_experiments
 MITgcm_download
 MITgcm_path
+verification_experiments
+setup_verification!
 testreport
 ```
