+++ src/ModelConfigurations.jl
@@ -17,6 +17,9 @@ MC=MITgcm_config(inputs=params,folder=folder)
 #providing executable (optional)
 #push!(MC.inputs[:setup][:main],(:exe => "./mitgcmuv"))
 
+#providing input folder (optional)
+#push!(MC.inputs[:setup][:main],(:input_folder => "./input_folder"))
+
 #modifying run time options (optional)
 #MC.inputs[:pkg][:PACKAGES][:useECCO]=false
 
@@ -31,32 +34,41 @@ launch(MC)
 ```
 """
 function setup_ECCO4!(config::MITgcm_config)
-    if !haskey(config.inputs[:setup],:build)
-        if !haskey(config.inputs[:setup][:main],:exe)
-            println("downloading MITgcm ... ")
-            u0="https://github.com/MITgcm/MITgcm"; p0=joinpath(config,"MITgcm")
-            @suppress run(`$(git()) clone --depth 1 --branch checkpoint68o $(u0) $(p0)`)
-            println("downloading code folder ... ")
-            u0="https://github.com/gaelforget/ECCOv4"; p0=joinpath(config,"ECCOv4")
-            @suppress run(`$(git()) clone $(u0) $(p0)`)
-            p1=joinpath(config,"MITgcm","mysetups")
-            p2=joinpath(p1,"ECCOv4")
-            mkdir(p1); mv(p0,p2)
-            p3=joinpath(p2,"build")
-            n3="mitgcmuv"
-        else
-            p3=dirname(config.inputs[:setup][:main][:exe])
-            n3=basename(config.inputs[:setup][:main][:exe])
-        end
-        P=OrderedDict(:path=>p3,:options=>"-mods=../code -mpi",:exe=>n3)
-        push!(config.inputs[:setup],(:build => P))
-        #push!(config.inputs[:setup][:main],(:command => "mpirun -np 96 mitgcmuv"))
-        push!(config.inputs[:setup][:main],(:command => "qsub job.csh"))
-        println("creating job submission script ...")
-        p=joinpath(pathof(config),"run")
-        f=joinpath(p,"submit.csh")
-        create_script(p,f)
+    if !haskey(config.inputs[:setup][:main],:exe)
+        println("downloading MITgcm ... ")
+        u0="https://github.com/MITgcm/MITgcm"; p0=joinpath(config,"MITgcm")
+        @suppress run(`$(git()) clone --depth 1 --branch checkpoint68o $(u0) $(p0)`)
+        println("downloading code folder ... ")
+        u0="https://github.com/gaelforget/ECCOv4"; p0=joinpath(config,"ECCOv4")
+        @suppress run(`$(git()) clone $(u0) $(p0)`)
+        p1=joinpath(config,"MITgcm","mysetups")
+        p2=joinpath(p1,"ECCOv4")
+        mkdir(p1); mv(p0,p2)
+        p3=joinpath(p2,"build")
+        n3="mitgcmuv"
+    else
+        p3=dirname(config.inputs[:setup][:main][:exe])
+        n3=basename(config.inputs[:setup][:main][:exe])
     end
+    P=OrderedDict(:path=>p3,:options=>"-mods=../code -mpi",:exe=>n3)
+    push!(config.inputs[:setup],(:build => P))
+    #push!(config.inputs[:setup][:main],(:command => "mpirun -np 96 mitgcmuv"))
+    push!(config.inputs[:setup][:main],(:command => "qsub job.csh"))
+    println("creating job submission script ...")
+    p=joinpath(pathof(config),"run")
[…]
