### 2024-05-15 e2e7564
fix test, skip download
