+++ docs/src/functionalities_read.md
@@ -45,6 +45,8 @@ read_nctiles
 
 Run-time parameters to `MITgcm` are provided via text files. This package can read and write them in two formats (standard `TOML` format, or the native `MITgcm_namelist` format).
 
+Other inputs (binary or NetCDF files) for gridded data, forcing fields, etc can be provided via an `input_folder` or downloaded as shown in the [`setup_ECCO4!`](@ref).
+
 ```@docs
 read_toml
 MITgcm_namelist
