### 2024-05-15 246fa76
add monitor, fixes
