+++ src/ModelConfigurations.jl
@@ -53,7 +53,7 @@ function setup_ECCO4!(config::MITgcm_config)
     P=OrderedDict(:path=>p3,:options=>"-mods=../code -mpi",:exe=>n3)
     push!(config.inputs[:setup],(:build => P))
     #push!(config.inputs[:setup][:main],(:command => "mpirun -np 96 mitgcmuv"))
-    push!(config.inputs[:setup][:main],(:command => "qsub job.csh"))
+    push!(config.inputs[:setup][:main],(:command => "qsub submit.csh"))
     println("creating job submission script ...")
     p=joinpath(pathof(config),"run")
     f=joinpath(p,"submit.csh")
