### 2024-06-14 gaelforget/MITgcm.jl#106 @simone-silvestri
[JOSS Review] Repository
1. I was able to run the example provided in https://gaelforget.github.io/MITgcm.jl/dev/examples/ on a Linux computer but not on my Mac M1. After some investigation, most likely it was a problem with the optfile for Apple silicon that was recently changed (probably after https://github.com/MITgcm/MITgcm/pull/613). Eventually, I was able to run the `advect_xy` configuration by specifying a NETCDF_ROOT directory. However, the error message was not instructive so it was difficult to assess the problem. I would suggest more verbose error messages such that the user can diagnose what is wrong (or even report the error through an issue on this repo)
2. A couple of bugs in the `MITgcm_scan_output.jl` example:
- When selecting the ECCO grid the error is `UndefVarError: Point2 not defined`
- When selecting the verification grid I get `type Missing has no field params_files` (I think it refers to the variable sc)
3. `MITgcm.jl_run` example: the only case that runs is the `advect_xy`, all the other configurations fail to build with the error `model compilation may have failed`
4. MITgcm_workflow.jl` example:
- The `advect_xy` configuration runs and shows a visualization of the results.
- The other cases fail with the error message `KeyError: key :setup not found`
6. I have problems with building other examples such as `HS94_animation.jl`
7. Given that most of the examples do not work on my laptop, I think it might be a configuration issue from my side. Providing a non-pluto example that produces a headless visualization (such as with CairoMakie) would help me test the repo on a remote linux machine.
8. as written in #105 (and given point 7.) I would add an example that runs in the terminal (for non-GUI people like myself) so that it is not necessary to open Pluto which requires a graphical user interface
