+++ ext/MITgcmNetCDFExt.jl
@@ -37,9 +37,9 @@ function read_nctiles(fileName::String,fldName::String,mygrid::gcmgrid;
 
     fileIn=joinpath(pth1,lst[1])
     fileRoot= eccoVersion4Release4 ? fileIn[1:end-11] : fileIn[1:end-8]
-    ntile=ClimateModels.ncgetatt(fileIn,"Global","ntile")
+    ntile=ncgetatt(fileIn,"Global","ntile")
 
-    x = ClimateModels.ncread(fileIn,fldName)
+    x = ncread(fileIn,fldName)
     s = [size(x,i) for i in 1:ndims(x)]
     n=length(size(x))
     start=ones(Int,n)
@@ -71,7 +71,7 @@ function read_nctiles(fileName::String,fldName::String,mygrid::gcmgrid;
                 fileCounter += 1
                 verbose ? @info("Reading file $(fileIn)") : nothing
                 for l in 1:13, k in 1:50
-                    x = ClimateModels.ncread(fileIn,fldName,start,count)
+                    x = ncread(fileIn,fldName,start,count)
                     face = tiles[l].face
                     i=collect(tiles[l].i)
                     j=collect(tiles[l].j)
@@ -95,7 +95,7 @@ function read_nctiles(fileName::String,fldName::String,mygrid::gcmgrid;
                 fileIn=@sprintf("%s.%04d.nc",fileRoot,n+n0)
                 if isfile(fileIn) #skip if no file / blank tile
                     verbose ? @info("Reading file $(fileIn)") : nothing
-                    x = ClimateModels.ncread(fileIn,fldName,start,count)
+                    x = ncread(fileIn,fldName,start,count)
                     i=collect(1:s[1]) .+ i0[n]
                     j=collect(1:s[2]) .+ j0[n]
                     f0[i,j,:,:]=x[:,:,:,:]
