### 2024-06-15 863ecac
update to use MITgcm-checkpoint68y-small.tar.gz and verif
