+++ src/ModelConfigurations.jl
@@ -646,6 +646,8 @@ path = ""
 function download_dataset(url,path)
     fname = joinpath(path, basename(url))
     if !isfile(fname)
+        !isdir(path) ? mkdir(path) : nothing
+#        println.(("b",url, fname))
         Downloads.download(url, fname)
     end
     return fname
