### 2024-06-15 9913624
symlink file to output.txt
