### 2024-06-15 a29a012
bump version to 0.4.0 to reflect update to latest MITgcm source code
