### 2024-06-19 gaelforget/MITgcm.jl#106 @gaelforget
> 1. I was able to run the example provided in https://gaelforget.github.io/MITgcm.jl/dev/examples/ on a Linux computer but not on my Mac M1. After some investigation, most likely it was a problem with the optfile for Apple silicon that was recently changed (probably after [new build-options file for darwin arm64 MITgcm/MITgcm#613](https://github.com/MITgcm/MITgcm/pull/613)).

Sorry for this! I had not updated the default version of MITgcm that MITgcm.jl uses. Should work now with MITgcm.jl v0.4.0 that uses MITgcm-checkpoint68.

> Eventually, I was able to run the `advect_xy` configuration by specifying a NETCDF_ROOT directory. However, the error message was not instructive so it was difficult to assess the problem. I would suggest more verbose error messages such that the user can diagnose what is wrong (or even report the error through an issue on this repo)

The NETCDF_ROOT issue seems to come from MITgcm itself, and maybe there's a bug with the genmake2 script that creates the Makefile.

Found these two issue : https://github.com/MITgcm/MITgcm/issues/82 & https://github.com/MITgcm/MITgcm/issues/635 . I believe there was some explanation about NETCDF_ROOT in the documentation but strangely that seems to have been cut, making things worse.

Like you, I find that I need NETCDF_ROOT defined to build the verification examples. But even if I do (line below that matches what I see in `nc-config --all`) then genmake2 turns off mnc (the package that would generated netcdf files). The examples run but generate binary output.

`export NETCDF_ROOT='/opt/homebrew/Cellar/netcdf/4.9.2_1/'

And If undefined NETCDF_ROOT then we end up with an incorrect Makefile that has :

`LIBS = -L/opt/homebrew/Cellar/netcdf-fortran/4.6.1/lib -lnetcdff -lnetcdf -lnetcdf`
