### 2024-06-19 gaelforget/MITgcm.jl#106 @simone-silvestri
> Sorry for this! I had not updated the default version of MITgcm that MITgcm.jl uses. Should work now with MITgcm.jl v0.4.0 that uses MITgcm-checkpoint68.

Thanks, then I'll try again after updating the repo and setting the `NETCDF_ROOT` flag
