### 2024-06-19 429d8ac
update dependencies + fix v Point2
