+++ examples/MITgcm_configurations.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.40
+# v0.19.42
 
 using Markdown
 using InteractiveUtils
@@ -12,12 +12,6 @@ macro bind(def, element)
         global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
         el
     end
-end
-
-# ╔═╡ 8cf4d8ca-84eb-11eb-22d2-255ce7237090
-begin
-	using MITgcm, PlutoUI, Printf
-	exps=verification_experiments()
 end
 
 # ╔═╡ f588eaba-84ef-11eb-0755-bf1b85b2b561
@@ -52,29 +46,6 @@ Each configuration has various groups of parameters, often called namelists, dep
 The chosen defaults, `data` and `PARM01`, are expected to be found in any MITgcm run directory. However, there can be many more parameters. 
 """
 
-# ╔═╡ f40e76c4-86d5-11eb-15b0-cd55d6cd1e65
-md"""### Appendices
-
-The following cells select Julia packages and perform basic operations.
-"""
-
-# ╔═╡ 168e178c-dd09-4e27-8cb6-fc0479a55f75
-begin
-	🏁 = "🏁"
-	imgA="https://user-images.githubusercontent.com/20276764/111042787-12377e00-840d-11eb-8ddb-64cc1cfd57fd.png"
-	imgB="https://user-images.githubusercontent.com/20276764/97648227-970b9780-1a2a-11eb-81c4-65ec2c87efc6.png"
-	md"""$(Resource(imgB, :width => 120))"""
-end
-
-# ╔═╡ 4965715d-93ca-496b-8ab1-238e9c6e34b4
-begin
-	iexp=findall([exps[i].configuration==myexp for i in 1:length(exps)])[1]
-	builddir=joinpath(MITgcm_path[1],"verification",myexp)
-	rundir=joinpath(exps[iexp].folder,string(exps[iexp].ID),"run")
-	!isdir(rundir) ? setup(exps[iexp]) : nothing
-	🏁
-end
-
 # ╔═╡ d7f2c656-8512-11eb-2fdf-47a3e57a55e6
 begin	
 	function list_namelist_files(pth)
@@ -92,13 +63,6 @@ begin
 	end
 end
 
-# ╔═╡ 348c692e-84fe-11eb-3288-dd0a1dedce90
-begin
-	fil=joinpath(rundir,mydats)
-	nml=read(fil,MITgcm_namelist())
-	🏁
-end
-
 # ╔═╡ ca7bb004-8510-11eb-379f-632c3b40723d
 try
 	@bind nmlgroup Select(String.(nml.groups))
@@ -113,17 +77,53 @@ md"""## Browse Parameters
 
 Now displaying 👉 **$myexp / $mydats : $nmlgroup**"""
 
+# ╔═╡ e50726aa-86d3-11eb-0418-fff8fb79ef95
+nml.params[inml]
+
+
+
+
+# ╔═╡ f40e76c4-86d5-11eb-15b0-cd55d6cd1e65
+md"""### Appendices
+
+The following cells select Julia packages and perform basic operations.
+"""
+
+# ╔═╡ 8cf4d8ca-84eb-11eb-22d2-255ce7237090
+begin
+	using MITgcm, PlutoUI, Printf
[…]
