### 2024-06-19 7990c90
rebuild and rerun by default
