### 2024-06-19 gaelforget/MITgcm.jl#106 @gaelforget
> 2 A couple of bugs in the MITgcm_scan_output.jl example:
When selecting the ECCO grid the error is UndefVarError: Point2 not defined
When selecting the verification grid I get type Missing has no field params_files (I think it refers to the variable sc)

Should be fixed after https://github.com/gaelforget/MITgcm.jl/pull/108 and https://github.com/gaelforget/MITgcm.jl/pull/109
