### 2024-06-19 gaelforget/MITgcm.jl#106 @gaelforget
> 3 MITgcm.jl_run example: the only case that runs is the advect_xy, all the other configurations fail to build with the error model compilation may have failed

Should be fixed (using v0.4.0 of MITgcm.jl + https://github.com/gaelforget/MITgcm.jl/pull/110)
