### 2024-06-19 gaelforget/MITgcm.jl#106 @gaelforget
> 4 MITgcm_workflow.jl` example:
The advect_xy configuration runs and shows a visualization of the results.
The other cases fail with the error message KeyError: key :setup not found

Should be fixed (using v0.4.0 of MITgcm.jl, NETCDF_ROOT -- to support ARM laptops)
