### 2024-06-19 gaelforget/MITgcm.jl#106 @gaelforget
> 7 Given that most of the examples do not work on my laptop, I think it might be a configuration issue from my side. Providing a non-pluto example that produces a headless visualization (such as with CairoMakie) would help me test the repo on a remote linux machine.
> 8 as written in https://github.com/gaelforget/MITgcm.jl/issues/105 (and given point 7.) I would add an example that runs in the terminal (for non-GUI people like myself) so that it is not necessary to open Pluto which requires a graphical user interface

Will address this @ https://github.com/gaelforget/MITgcm.jl/issues/105

Good points!
