+++ test/runtests.jl
@@ -1,4 +1,4 @@
-using MITgcm, NetCDF, OceanStateEstimation, Test
+using MITgcm, NetCDF, Climatology, Test
 
 using MITgcm.MeshArrays
 using MITgcm.ClimateModels.Suppressor
