### 2024-06-20 #6710 @AnjaliSandip
👋 @haakon-e, @briochemc, @Alexander-Barth, @visr, @barbara42, @glwagner, @navidcy, @ jbisits, @swilliamson7, @fonsp, @SimonDanisch, @Datseris, @szy21, and @LenkaNovak, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
