### 2024-06-20 #6710 @Datseris
I actually wanted to run MITGCM some time ago, but I couldn't be bothered to learn fortran interfaces. This comes as a pleasant surprise. Sure, I'll serve as a reviewer.
