### 2024-06-20 #6710 @simone-silvestri
> @simone-silvestri How close are you to completing the review?

I submitted my review as issues in the MITgcm.jl repository. Since my review has already been addressed, I need to verify that everything works correctly and then I can consider it complete
