### 2024-06-21 gaelforget/MITgcm.jl#106 @gaelforget
> 7 Given that most of the examples do not work on my laptop,

Hi @simone-silvestri

About building on your laptop, can you try with `export NETCDF_ROOT=/opt/homebrew` ?

This fixed it for me.

See jahn's suggestion @ https://github.com/MITgcm/MITgcm/issues/846#issuecomment-2181691225
