### 2024-06-21 c78cc1c
add MITgcm.set_environment_variables_to_default
