+++ docs/src/functionalities_configurations.md
@@ -78,4 +78,5 @@ MITgcm_path
 verification_experiments
 setup_verification!
 testreport
+MITgcm.set_environment_variables_to_default
 ```
