### 2024-06-21 gaelforget/MITgcm.jl#106 @gaelforget
Great! Thank you

ps. I am about to release v0.4.1

- includes ```MITgcm.set_environment_variables_to_default``` as a basic way to define `NETCDF_ROOT` and `MPI_INC_DIR`
- to be improved later ...
