### 2024-06-21 027eb77
add build option presets
