### 2024-06-21 41e2468
new MITgcm.default_path() that downloads MITgcm if needed
