+++ docs/src/functionalities_more.md
@@ -1,5 +1,12 @@
 # Additional Functionalities
 
+### Related packages 
+
+- [ClimateModels.jl](https://gaelforget.github.io/ClimateModels.jl/stable/) : standard interface for modeling workflows. 
+- [OceanStateEstimation.jl](https://JuliaOcean.github.io/OceanStateEstimation.jl/dev/) : accessing and using ECCO solutions. 
+- [MeshArrays.jl](https://juliaclimate.github.io/MeshArrays.jl/dev/) : data structures for gridded model output.
+- [IndividualDisplacements.jl](https://juliaclimate.github.io/IndividualDisplacements.jl/dev/) : particle tracking applications and pathway simulationss. 
+
 ### Format Conversions
 
 ```@docs
