+++ docs/src/index.md
@@ -13,25 +13,17 @@ Pages = [
     "examples.md",
     "contributing.md",
 ]
-Depth = 2
+Depth = 1
 ```
 
 ## Main Features
 
 - Read / Write / Scan the various MITgcm file types
-- Standard Modeling Worflow (`ClimateModels`' interface)
-- Standard MITgcm configurations (`ECCO`, `verification`)
-- [Examples](@ref) (docs, and notebooks in `examples/`)
+- Standard MITgcm Run Workflow (Julia interface)
+- Standard MITgcm configurations (`verification`, `ECCO`)
+- Examples (docs, and notebooks in `examples/`)
 
-## Related packages 
-
-- [ClimateModels.jl](https://gaelforget.github.io/ClimateModels.jl/stable/) : standard interface for modeling workflows. 
-- [OceanStateEstimation.jl](https://JuliaOcean.github.io/OceanStateEstimation.jl/dev/) : accessing and using ECCO solutions. 
-- [MeshArrays.jl](https://juliaclimate.github.io/MeshArrays.jl/dev/) : data structures for gridded model output.
-- [IndividualDisplacements.jl](https://juliaclimate.github.io/IndividualDisplacements.jl/dev/) : particle tracking applications and pathway simulationss. 
-
-## Visuals 
-
-![Simulated particles from HS94 on cube sphere grid](https://user-images.githubusercontent.com/20276764/111042787-12377e00-840d-11eb-8ddb-64cc1cfd57fd.png)
-
-![The impossible MITgcm rendering](https://user-images.githubusercontent.com/20276764/97648227-970b9780-1a2a-11eb-81c4-65ec2c87efc6.png)
+```@raw html
+<img src="https://user-images.githubusercontent.com/20276764/111042787-12377e00-840d-11eb-8ddb-64cc1cfd57fd.png" alt="Simulated particles from HS94 on cube sphere grid" width="200">
+<img src="https://user-images.githubusercontent.com/20276764/97648227-970b9780-1a2a-11eb-81c4-65ec2c87efc6.png" alt="The impossible MITgcm rendering" width="200">
+```
