+++ src/ReadFiles.jl
@@ -99,8 +99,13 @@ function scan_stdout(filout::String)
         ioSize[1]=size(tmp)
     end
     if tst_mnc
-        tmp=read_mnc(pth_mnc,"grid","XC")
-        ioSize[1]=size(tmp)
+        try
+            tmp=read_mnc(pth_mnc,"grid","XC")
+            ioSize[1]=size(tmp)
+        catch e
+            @warn "failed to read mnc file (using NetCDF ?)"
+            ioSize[1]=(0,0)
+        end
     end
 
     par3=(use_mdsio=tst_mdsio,use_mnc=tst_mnc,ioSize=ioSize[1])
@@ -364,8 +369,8 @@ end
 Read all `MITgcm` namelist files in `input_path`, parse them, and return as a NamedTuple of NamedTuples.
 
 ```
-using MITgcm; MITgcm_download()
-input_path=joinpath(MITgcm_path[1],"verification","advect_xy","input")
+using MITgcm; path0=default_path()
+input_path=joinpath(path0,"verification","advect_xy","input")
 params=read_all_namelists(input_path)
 ```
 """
