+++ docs/make.jl
@@ -6,7 +6,7 @@ makedocs(;
     modules=[MITgcm,Base.get_extension(MITgcm, :MITgcmNetCDFExt)],
     format=Documenter.HTML(),
     pages=[
-        "Home" => "index.md",
+        "Contents" => "index.md",
         "Manual" => ["functionalities.md","functionalities_configurations.md",
         "functionalities_interface.md","functionalities_read.md","functionalities_more.md"],
         "Examples" => "examples.md",
