+++ docs/src/functionalities_configurations.md
@@ -73,4 +73,5 @@ verification_experiments
 setup_verification!
 testreport
 MITgcm.set_environment_variables_to_default
+MITgcm.system_check
 ```
