+++ docs/src/functionalities_interface.md
@@ -1,10 +1,11 @@
 # Interface to MITgcm
 
-The [ClimateModels.jl](https://github.com/gaelforget/ClimateModels.jl/#readme)'s interface is implemented in `MITgcm.jl`. This allows you to easily create model simulations using `MITgcm` conveniently from `Julia`.
+The [ClimateModels.jl](https://github.com/gaelforget/ClimateModels.jl/#readme)'s interface is implemented in `MITgcm.jl`. 
+
+This allows you to easily create model simulations using `MITgcm` conveniently from `Julia`.
 
 ```@example 1
 using MITgcm
-MITgcm.set_environment_variables_to_default()
 MC=MITgcm_config(configuration="advect_xy")
 setup(MC)
 build(MC,"--allow-skip")
