+++ docs/src/index.md
@@ -7,8 +7,8 @@ Julia interface to [MITgcm](https://mitgcm.readthedocs.io/en/latest/?badge=lates
 ```@contents
 Pages = [
     "functionalities.md",
-    "functionalities_configurations.md",
     "functionalities_interface.md",
+    "functionalities_configurations.md",
     "functionalities_read.md",
     "examples.md",
     "contributing.md",
