### 2024-06-23 055c36e
do_warn=false by default + docs improvement
