+++ src/MITgcm.jl
@@ -27,9 +27,10 @@ export monitor #pause, stop, clock, train, help
 export verification_experiments, setup_verification!, testreport
 export setup_ECCO4!, ECCO4_inputs, ECCO4_testreport
 export read_namelist, write_namelist, read_toml
-export read_all_namelists, write_all_namelists
+export read_all_namelists, write_all_namelists, parse_param
 export read_mdsio, read_meta, read_available_diagnostics
-export scan_rundir, scan_stdout
+export scan_rundir, scan_stdout, create_script, default_path
+export set_environment_variables_to_default, system_check
 export read_bin, read_flt, read_mnc, read_nctiles, findtiles
 export GridLoad_mnc, GridLoad_mdsio, GridLoad_native
 export cube2compact, compact2cube, convert2array, convert2gcmfaces
