+++ docs/src/functionalities_interface.md
@@ -8,21 +8,36 @@ This allows you to easily create model simulations using `MITgcm` conveniently f
 using MITgcm
 MC=MITgcm_config(configuration="advect_xy")
 setup(MC)
-build(MC,"--allow-skip")
+exe=joinpath(MITgcm.default_path(),"verification",MC.configuration,"build","mitgcmuv") #hide
+MC.inputs[:setup][:build][:exe]=exe #hide
+build(MC)
 launch(MC)
 ```
 
 ```@example 1
-sc=scan_rundir(joinpath(MC,"run"))
+sc=scan_rundir(MC)
 keys(sc)
 ```
 
 ## Functionalities
 
+### Main
+
 ```@docs
 setup
 build
-compile
 MITgcm_launch
-clean
 ```
+
+### Tools
+
+```@docs
+system_check
+set_environment_variables_to_default
+default_path
+MITgcm_download
+create_script
+parse_param
+compile
+clean
+```│
