### 2024-06-23 b06453d
skip mnc part
