+++ docs/src/functionalities_read.md
@@ -13,6 +13,21 @@ scan_stdout
 read_available_diagnostics
 ```
 
+## Input Files
+
+Run-time parameters to `MITgcm` are provided via text files. This package can read and write them in two formats (standard `TOML` format, or the native `MITgcm_namelist` format).
+
+Other inputs (binary or NetCDF files) for gridded data, forcing fields, etc can be provided via an `input_folder` or downloaded as shown in the [`setup_ECCO4!`](@ref).
+
+```@docs
+read_toml
+MITgcm_namelist
+read_namelist
+write_namelist
+read_all_namelists
+write_all_namelists
+```
+
 ## MDS Files
 
 ```@docs
@@ -42,17 +57,3 @@ read_bin
 read_nctiles
 ```
 
-## Input Files
-
-Run-time parameters to `MITgcm` are provided via text files. This package can read and write them in two formats (standard `TOML` format, or the native `MITgcm_namelist` format).
-
-Other inputs (binary or NetCDF files) for gridded data, forcing fields, etc can be provided via an `input_folder` or downloaded as shown in the [`setup_ECCO4!`](@ref).
-
-```@docs
-read_toml
-MITgcm_namelist
-read_namelist
-write_namelist
-read_all_namelists
-write_all_namelists
-```
