### 2024-06-24 7f98771
add code snippet example + mv plots + add funding refs
