### 2024-06-25 gaelforget/MITgcm.jl#106 @gaelforget
> Given that most of the examples do not work on my laptop, I think it might be a configuration issue from my side. Providing a non-pluto example that produces a headless visualization (such as with CairoMakie) would help me test the repo on a remote linux machine.

done

ps. about _configuration issue from my side_, I had the same issue. see https://github.com/MITgcm/MITgcm/issues/846

> as written in https://github.com/gaelforget/MITgcm.jl/issues/105 (and given point 7.) I would add an example that runs in the terminal (for non-GUI people like myself) so that it is not necessary to open Pluto which requires a graphical user interface

done

ps. I don't think that running Pluto notebooks requires a graphical user interface. In a terminal window, or batch, `using Pluto; Pluto.activate_notebook_environment("nb.jl"); include(nb.jl)` as well as `PlutoSliderServer.export_notebook("nb.jl") should work.
