### 2024-06-25 gaelforget/MITgcm.jl#106 @gaelforget
All set of this issue (1 to 8)?
