+++ examples/Darwin3_1D.jl
@@ -1,0 +1,1806 @@
+### A Pluto.jl notebook ###
+# v0.19.43
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ 95380a98-b33d-48b2-b244-3e82509c2198
+begin
+	using MITgcm
+	using PlutoUI, CairoMakie
+end
+
+# ╔═╡ 91de38ed-9b1d-4b91-8744-1e0db6c4a60d
+md"""# Darwin3 Model (1D)
+
+**Contributors:**
+
+- Stephanie Dutkiewicz provided the examples
+- Oliver Jahn and collaborators the Darwin3 source code
+- Gaël Forget the MITgcm.jl interface and notebook prototype
+
+!!! note
+	For more information, please refer to 
+	- [darwin3.readthedocs](https://darwin3.readthedocs.io/en/latest/phys_pkgs/darwin.html)
+	- [darwinpropject.mit.edu](https://darwinproject.mit.edu)
+	- [cbiomes.org](https://darwinproject.mit.edu)
+	- [MITgcm.readthedocs](http://mitgcm.readthedocs.io/en/latest/?badge=latest)
+	- [MITgcm.jl](https://github.com/gaelforget/MITgcm.jl#readme).
+"""
+
+# ╔═╡ cbda8aa0-c483-4038-a6dc-957f2e6f8c7b
+begin
+	ptr_select = @bind ptr Select(1:10)
+	md"""## Model Output
+
+	ptracer ID = $(ptr_select)
+	"""
+end
+
+# ╔═╡ ac48d134-6eaf-4df8-a476-3a06fe035135
+md"""## Model Parameters"""
+
+# ╔═╡ 2d88fc56-ec4c-48cd-92ae-9dd461b2d9fd
+md"""## Model Run"""
+
+# ╔═╡ 301b38d9-a3a0-4f6b-9892-71f86932d1c1
+begin
+	MC=MITgcm_config(model="darwin3",configuration="31+16+3_RT_1D")
+	setup(MC)
+
+    MC.inputs[:pkg][:PACKAGES][:useMNC]=false
+    write_all_namelists(MC.inputs,joinpath(MC,"run"))
+    
+	build(MC)
+	MITgcm_launch(MC)
+end
+
+# ╔═╡ b12b379b-02d3-477e-b720-e031355fcfc2
+begin	
+	RC=read_mdsio(joinpath(MC,"run","RC.data"))[:]
+	msk=read_mdsio(joinpath(MC,"run","hFacC.data"))[:] 
+	msk=[(m>0 ? 1 : NaN) for m in msk]
+	tracer=read_mdsio(joinpath(MC,"run","PTRACER01.0000000000.data"))[:]
[…]
