+++ docs/src/functionalities_interface.md
@@ -8,7 +8,7 @@ This allows you to easily create model simulations using `MITgcm` conveniently f
 using MITgcm
 MC=MITgcm_config(configuration="advect_xy")
 setup(MC)
-exe=joinpath(MITgcm.default_path(),"verification",MC.configuration,"build","mitgcmuv") #hide
+exe=joinpath(MITgcm_path[2],MC.configuration,"build","mitgcmuv") #hide
 MC.inputs[:setup][:build][:exe]=exe #hide
 build(MC)
 launch(MC)
