+++ examples/Darwin3_1D.jl
@@ -70,14 +70,19 @@ begin
 	RC=read_mdsio(joinpath(MC,"run","RC.data"))[:]
 	msk=read_mdsio(joinpath(MC,"run","hFacC.data"))[:] 
 	msk=[(m>0 ? 1 : NaN) for m in msk]
-	tracer=read_mdsio(joinpath(MC,"run","PTRACER01.0000000000.data"))[:]
+	qtr=(ptr>9 ? "$(ptr)" : "0$(ptr)")
+	tracer_0=read_mdsio(joinpath(MC,"run","PTRACER$(qtr).0000000000.data"))[:]
+	tracer_64=read_mdsio(joinpath(MC,"run","PTRACER$(qtr).0000000064.data"))[:]
 end
+
+# ╔═╡ 2d2f310f-9f67-46a6-a0b7-cf96e2268f39
+readdir(MC,"run")
 
 # ╔═╡ 854fc514-8b7d-47d9-9bb1-7fae4528e020
 MC.inputs
 
 # ╔═╡ ac880e26-48b1-4001-a84b-1f624e746639
-md"""## appendix"""
+md"""## Appendix"""
 
 # ╔═╡ 7bb2774f-e23d-488e-a58b-bcd530856708
 function ptr_name(ptr)
@@ -88,8 +93,9 @@ end
 
 # ╔═╡ 9b6a364a-48b6-4c93-8707-4c32d1f8759a
 begin
-	fig,ax,li=lines(msk.*tracer,RC)
-	ax.title=ptr_name(ptr)
+	fig,ax,li=lines(msk.*tracer_0,RC,label="time = 0")
+	lines!(msk.*tracer_64,RC,label="time = 64 steps")
+	ax.title=ptr_name(ptr); axislegend(ax,position=:lb)
 	fig
 end
 
@@ -102,7 +108,7 @@ PlutoUI = "7f904dfe-b85e-4ff6-b463-dae2292396a8"
 
 [compat]
 CairoMakie = "~0.12.3"
-MITgcm = "~0.4.1"
+MITgcm = "~0.4.4"
 PlutoUI = "~0.7.59"
 """
 
@@ -112,7 +118,7 @@ PLUTO_MANIFEST_TOML_CONTENTS = """
 
 julia_version = "1.10.4"
 manifest_format = "2.0"
-project_hash = "0c6665110c7d9523668e1aa414e31bf19846e7ed"
+project_hash = "3c2837f5bf9b4345f87cd6027297a917b88b73fe"
 
 [[deps.AbstractFFTs]]
 deps = ["LinearAlgebra"]
@@ -359,6 +365,12 @@ version = "4.1.1"
 git-tree-sha1 = "abe83f3a2f1b857aac70ef8b269080af17764bbe"
 uuid = "9a962f9c-6df0-11e9-0e5d-c546b8b5ee8a"
 version = "1.16.0"
+
+[[deps.DataDeps]]
+deps = ["HTTP", "Libdl", "Reexport", "SHA", "Scratch", "p7zip_jll"]
+git-tree-sha1 = "8ae085b71c462c2cb1cfedcb10c3c877ec6cf03f"
+uuid = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
+version = "0.7.13"
 
 [[deps.DataFrames]]
[…]
