+++ v0.4.4/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"functionalities/#Overview","page":"Overview","title":"Overview","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"MITgcm.jl allows the analysis of MITgcm results in Julia. It can also setup, build, and launch a chosen model configuration (MITgcm_config) from within julia. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Functionalities are documented in the coming sections, and in Examples, Notebooks.","category":"page"},{"location":"functionalities/#Getting-Started","page":"Overview","title":"Getting Started","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Installing the latest version of MITgcm.jl with the built-in package manager is the recommended method. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"using Pkg\nPkg.add(\"MITgcm\")","category":"page"},{"location":"functionalities/#Running-MITgcm","page":"Overview","title":"Running MITgcm","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Let's start running MITgcm interactively. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"The first command defines a data structure, MITgcm_config, that we can then use from Julia. The setup command, by default, creates a folder in your tempdir() to run MITgcm.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"using MITgcm\nMC=MITgcm_config(configuration=\"tutorial_held_suarez_cs\")\nsetup(MC)\nexe=joinpath(MITgcm.default_path(),\"verification\",MC.configuration,\"build\",\"mitgcmuv\") #hide\nshow(MC)","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"After setup, the standard workflow is to call build and then MITgcm_launch. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"Next we can use the log method to get a status report. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"build(MC)\nlaunch(MC)\nlog(MC)","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"If we have a previous build of MITgcm for this then we can specify the file path as a parameter, :exe. ","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"If :exe is not specified, or if no file is found at the specified path, build will attempt to build the model for us.","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"exe=joinpath(default_path(),\"verification\",MC.configuration,\"build\",\"mitgcmuv\")\nMC.inputs[:setup][:build][:exe]=exe","category":"page"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"tip: Tip\nFor longer MITgcm simulations run, users often prefer to use a queuing system or batch script (not an interactive session).\nsetup can generate and submit a batch script, via create_script and config.inputs[:setup][:main][:command] = \"qsub submit.csh\"","category":"page"},{"location":"functionalities/#Using-Model-Output","page":"Overview","title":"Using Model Output","text":"","category":"section"},{"location":"functionalities/","page":"Overview","title":"Overview","text":"As MITgcm users, we often want to read and visualise model output from an earl
[…]
