### 2024-06-30 9791d43
update notebook deps
