### 2024-07-01 161d7b9
force downloads for Darwin3 example
