### 2024-07-12 gaelforget/MITgcm.jl#106 @gaelforget
Hi @simone-silvestri

quick bump. I should be able to work on this next week if needed, unless we are ready to close this one?
