### 2024-07-12 #6710 @gaelforget
Hi @AnjaliSandip

-  regarding reviewer 1 , we might be nearing completion (https://github.com/gaelforget/MITgcm.jl/issues/106).
- regarding reviewer 2 , do we need to create the checklist for them ? not sure
