### 2024-07-20 523358e
heat budget code, first draft
