+++ examples/budgets/modules.jl
@@ -1,0 +1,566 @@
+
+##
+
+module GRID
+  using MeshArrays, JLD2
+
+  γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
+
+  τ=Tiles(γ,30,30)
+
+  import Base: zeros
+  function zeros(g::gcmgrid)
+    tmp=MeshArray(γ)
+    tmp.=0
+    tmp
+  end
+
+  lon=GridLoadVar("XC",γ)
+  lat=GridLoadVar("YC",γ)
+
+  hFacC=GridLoadVar("hFacC",γ)
+  M=GridLoadVar("hFacC",γ)
+  [M[:,k]=land_mask(M[:,k]) for k in 1:50]
+
+  dep=-GridLoadVar("RC",γ)
+  dep_RF=-GridLoadVar("RF",γ)
+  area=GridLoadVar("RAC",γ)
+
+  DXG=GridLoadVar("DXG",γ)
+  DYG=GridLoadVar("DYG",γ)
+
+  DRF=GridLoadVar("DRF",γ)
+  RAC=GridLoadVar("RAC",γ)
+  Depth=GridLoadVar("Depth",γ)
+
+  nr=length(dep)
+
+  rhoconst =1029; #sea water density
+  rcp      =3994*rhoconst; # sea water rho X heat capacity
+
+  dlat=2
+  lats=(-90+dlat/2:dlat:90-dlat/2)
+  nl=length(lats)
+
+### basins
+
+basins=demo.ocean_basins()
+
+list_Atl=["Atlantic","Gulf of Mexico","Hudson Bay","Mediterranean Sea","North Sea","Baffin Bay","GIN Seas"]
+list_Pac=["Pacific","Bering Sea","Okhotsk Sea","Japan Sea","East China Sea"]
+list_Ind=["indian","South China Sea","Java Sea","Timor Sea","Red Sea","Gulf"]
+list_Arc=["Arctic","Barents Sea"]
+
+function larger_basins(list_Atl)
+  map_Atl=0*basins.map
+  for i in list_Atl
+   #println(i)
+   jj=findall(basins.name.==i)[1]
+   map_Atl.=map_Atl+1.0*(basins.map.==jj)
+  end
+  map_Atl
+end
+
+map_Atl=larger_basins(list_Atl)
+map_Pac=larger_basins(list_Pac)
+map_Ind=larger_basins(list_Ind)
+map_Arc=larger_basins(list_Arc)
+
+end
+
+##
+
+module IO_CLIM
+
+using JLD2, Glob, MITgcm, MeshArrays, Statistics
+
+import MITgcm: read_mdsio
+read_mdsio(γ::gcmgrid,args...;kwargs...) = read(read_mdsio(args...;kwargs...),γ)
+
+###
+
+diags_path="OCCA2HR1/run2"
+stat_path="dT_run_v5_stat_v5"
+
+files=glob("diags/budg*.0000000732.meta",diags_path)
+isempty(files) ? files=glob("diags/budg*.0000000024.meta",diags_path) : nothing
+isempty(files) ? error("no files found") : nothing
+
+variables=[read_meta(f).fldList for f in files]
+files=[basename(split(f,".")[1]) for f in files]
+times=glob("diags/$(files[1])*data",diags_path)
+times=parse.(Float64,[split(f,".")[end-1] for f in times])
+
+γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
+
+###
+
[…]
