### 2024-07-31 32e56e4
update notebooks
