+++ examples/HS94_animation.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.43
+# v0.19.45
 
 using Markdown
 using InteractiveUtils
@@ -7,10 +7,11 @@ using InteractiveUtils
 # ╔═╡ 3668f786-9597-11eb-01a1-87d34b49eef9
 begin	
 	#packages for I/O, interpolation, etc	
-	using MITgcm, MeshArrays, Plots, PlutoUI
+	using MITgcm, DataDeps, MeshArrays, Plots, PlutoUI
 	withenv("DATADEPS_ALWAYS_ACCEPT"=>true) do
 		PICKUP_hs94_path=MITgcm.getdata("hs94pickup")
 	end		
+	PICKUP_hs94_path=MITgcm.getdata("hs94pickup")
 	🏁 = "🏁"
 	"Downloads and packages : complete."
 end
@@ -219,6 +220,7 @@ sc
 # ╔═╡ 00000000-0000-0000-0000-000000000001
 PLUTO_PROJECT_TOML_CONTENTS = """
 [deps]
+DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 MITgcm = "dce5fa8e-68ce-4431-a242-9469c69627a0"
 MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
@@ -231,7 +233,7 @@ PLUTO_MANIFEST_TOML_CONTENTS = """
 
 julia_version = "1.10.4"
 manifest_format = "2.0"
-project_hash = "f5511684011a42f99e2b72f13cb86f48a6622a6a"
+project_hash = "c8507d7b06440a7f5d8f484be36abff8f8677aa8"
 
 [[deps.AbstractPlutoDingetjes]]
 deps = ["Pkg"]
@@ -296,15 +298,15 @@ version = "0.3.5"
 
 [[deps.CodecZlib]]
 deps = ["TranscodingStreams", "Zlib_jll"]
-git-tree-sha1 = "59939d8a997469ee05c4b4944560a820f9ba0d73"
+git-tree-sha1 = "b8fe8546d52ca154ac556809e10c75e6e7430ac8"
 uuid = "944b1d66-785c-5afd-91f1-9de20f533193"
-version = "0.7.4"
+version = "0.7.5"
 
 [[deps.ColorSchemes]]
 deps = ["ColorTypes", "ColorVectorSpace", "Colors", "FixedPointNumbers", "PrecompileTools", "Random"]
-git-tree-sha1 = "4b270d6465eb21ae89b732182c20dc165f8bf9f2"
+git-tree-sha1 = "b5278586822443594ff615963b0c09755771b3e0"
 uuid = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
-version = "3.25.0"
+version = "3.26.0"
 
 [[deps.ColorTypes]]
 deps = ["FixedPointNumbers", "Random"]
@@ -347,9 +349,9 @@ version = "1.1.1+0"
 
 [[deps.ConcurrentUtilities]]
 deps = ["Serialization", "Sockets"]
-git-tree-sha1 = "6cbbd4d241d7e6579ab354737f4dd95ca43946e1"
+git-tree-sha1 = "ea32b83ca4fefa1768dc84e504cc0a94fb1ab8d1"
 uuid = "f0e56b4a-5159-44fe-b623-3e5288b988bb"
-version = "2.4.1"
+version = "2.4.2"
 
 [[deps.Contour]]
 git-tree-sha1 = "439e35b0b36e2e5881738abc8857bd92ad6ff9a8"
[…]
