+++ examples/HS94_particles.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.43
+# v0.19.45
 
 using Markdown
 using InteractiveUtils
@@ -136,6 +136,43 @@ md""" ### Trajectory Computation
 The particle trajectory output is in tabular format. 
 """
 
+# ╔═╡ 2d320d61-9e55-471e-b663-8d95f2055b84
+md"""### Save Results
+
+A simple yet general method is demonstrated in the next cell:
+"""
+
+# ╔═╡ 5129ac14-0861-40b5-b97e-0d0edd37849f
+md""" ### Plots
+
+The plots below show:
+
+- 1. initial positions (red) and after 1 day (blue; `x,y`)
+- 2. initial and final positions ($ndt days; `x,y`)
+- 3. initial and final coordinates ($ndt days; `lon,lat`)
+
+To save a plot, add somthing like `savefig(tempdir()*"/"*"hs94.cs.particles.png")`
+"""
+
+# ╔═╡ e9535b96-07c2-4aff-964d-bb327c5f735b
+begin
+	"""
+	    plot(𝐼::Individuals)
+	Plot the initial and final positions as scatter plot in x,y plane.
+	"""
+	function plot(𝐼::Individuals)
+	    🔴_by_t = IndividualDisplacements.groupby(𝐼.🔴, :t)
+	    if (sum(names(🔴_by_t).=="lon")==0)
+	        fig=scatter(🔴_by_t[1].x,🔴_by_t[1].y,c=:red,label="t0",marker = (:circle, stroke(0)))
+	        scatter!(🔴_by_t[end].x,🔴_by_t[end].y,c=:blue,label="t1",marker = (:circle, stroke(0)))
+	    else
+	        fig=scatter(🔴_by_t[1].lon,🔴_by_t[1].lat,c=:red,label="t0",marker = (:circle, stroke(0)))
+	        scatter!(🔴_by_t[end].lon,🔴_by_t[end].lat,c=:blue,label="t1",marker = (:circle, stroke(0)))
+	    end
+	    return fig
+	end
+end
+
 # ╔═╡ 7cc13ee0-ad61-4274-9c94-cbfa3974046b
 begin
 	#Read velocity fields and compute trajectories
@@ -158,12 +195,6 @@ begin
 	🔴 = 𝐼.🔴
 end
 
-# ╔═╡ 2d320d61-9e55-471e-b663-8d95f2055b84
-md"""### Save Results
-
-A simple yet general method is demonstrated in the next cell:
-"""
-
 # ╔═╡ d6ce6158-bdf0-48cd-8c6e-d6c1191e6a07
 begin
 	🔴 #using this variable for sync
@@ -173,18 +204,6 @@ begin
 	joinpath(PICKUP_hs94_path,"HS94.jld2")
 end
 
-# ╔═╡ 5129ac14-0861-40b5-b97e-0d0edd37849f
-md""" ### Plots
-
-The plots below show:
-
-- 1. initial positions (red) and after 1 day (blue; `x,y`)
-- 2. initial and final positions ($ndt days; `x,y`)
-- 3. initial and final coordinates ($ndt days; `lon,lat`)
-
-To save a plot, add somthing like `savefig(tempdir()*"/"*"hs94.cs.particles.png")`
-"""
-
 # ╔═╡ a34adf36-81de-42c2-8407-d408041193b6
 f1
 
@@ -193,25 +212,6 @@ f2
 
[…]
