### 2024-07-31 gaelforget/MITgcm.jl#106 @gaelforget
> 6 I have problems with building other examples such as HS94_animation.jl

Issue remained which should be fixed in https://github.com/gaelforget/MITgcm.jl/pull/123
