### 2024-08-01 gaelforget/MITgcm.jl#106 @simone-silvestri
Here are the output of running the examples of two different machines.
all the examples are run by setting
```julia
julia> using Pluto

julia> Pluto.activate_notebook_environment("examples/example.jl")

(example.jl) pkg> instantiate
```

# Macbook pro, M1

### MITgcm_configurations

Runs!

### MITgcm_workflow.jl

```Julia
julia> include("examples/MITgcm_worklow.jl")
ERROR: LoadError: KeyError: key :setup not found
Stacktrace:
 [1] getindex
   @ ~/.julia/packages/OrderedCollections/9C4Uz/src/ordered_dict.jl:385 [inlined]
 [2] build(config::MITgcm_config)
   @ MITgcm ~/.julia/packages/MITgcm/apnJ2/src/ModelSteps.jl:32
 [3] top-level scope
   @ ~/development/MITgcm.jl/examples/MITgcm_worklow.jl:89
 [4] include(fname::String)
   @ Base.MainInclude ./client.jl:489
 [5] top-level scope
   @ REPL[10]:1
in expression starting at /Users/loaner/development/MITgcm.jl/examples/MITgcm_worklow.jl:87
```

### MITgcm_run.jl

Failed for an unspecified reason

```julia
(MITgcm_run.jl) pkg> instantiate

julia> include("examples/MITgcm_run.jl")
model compilation may have failed
"# This file is machine-generated - editing it directly is not advised\n\njulia_version = \"1.10.4\"\nmanifest_format = \"2.0\"\nproject_hash = \"78ba067645459166fe1dc27b5aa25fa01a547b1c\"\n\n[[deps.AbstractPlutoDingetjes]]\ndeps = [\"Pkg\"]\ngit-tree-sha1 = \"6e1d2a35f2f90a4bc7c2ed98079b2ba09c35b83a\"\nuuid = \"6e696c72-6542-2067-7265-42206c756150\"\nversion = \"1.3.2\"\n\n[[deps.ArgTools]]\nuuid = \"0dad84c5-d112-42e6-8d28-ef12dabb789f\"\nversion = \"1.1.1\"\n\n[[deps.A" ⋯ 20614 bytes ⋯ "[\"Libdl\"]\nuuid = \"83775a58-1f1d-513f-b197-d71354ab007a\"\nversion = \"1.2.13+1\"\n\n[[deps.libblastrampoline_jll]]\ndeps = [\"Artifacts\", \"Libdl\"]\nuuid = \"8e850b90-86db-534c-a0d3-1478176c7d93\"\nversion = \"5.8.0+1\"\n\n[[deps.nghttp2_jll]]\ndeps = [\"Artifacts\", \"Libdl\"]\nuuid = \"8e850ede-7688-5339-a07c-302acd2aaf8d\"\nversion = \"1.52.0+1\"\n\n[[deps.p7zip_jll]]\ndeps = [\"Artifacts\", \"Libdl\"]\nuuid = \"3f19e933-33d8-53b3-aaab-bd5110c3b7a0\"\nversion = \"17.4.0+2\"\n"
```

### MITgcm_scan_output.jl

Runs!

### HS94_animation.jl

```julia
julia> include(“examples/HS94_animation.jl”)
ERROR: LoadError: UndefVarError: `PICKUP_hs94_path` not defined
Stacktrace:
 [1] top-level scope
   @ ~/development/MITgcm.jl/examples/HS94_animation.jl:78
 [2] include(fname::String)
   @ Base.MainInclude ./client.jl:489
 [3] top-level scope
   @ REPL[4]:1
in expression starting at /Users/loaner/development/MITgcm.jl/examples/HS94_animation.jl:78
```

### Other HS94 examples

Did not try them because contingent on `HS94_animation.jl` to complete

# Ubuntu 20.04.6 LTS

### MITgcm_configurations

Runs!

### MITgcm_workflow.jl

```julia
julia> include("examples/MITgcm_worklow.jl")
ERROR: LoadError: KeyError: key :setup not found
Stacktrace:
 [1] getindex
   @ ~/.julia/packages/OrderedCollections/9C4Uz/src/ordered_dict.jl:385 [inlined]
 [2] build(config::MITgcm_config)
   @ MITgcm ~/.julia/packages/MITgcm/apnJ2/src/ModelSteps.jl:32
 [3] top-level scope
   @ ~/development/MITgcm.jl/examples/MITgcm_worklow.jl:89
 [4] include(fname::String)
   @ Base.MainInclude ./client.jl:489
 [5] top-level scope
   @ REPL[10]:1
in expression starting at /Users/loaner/development/MITgcm.jl/examples/MITgcm_worklow.jl:87
```

### MITgcm_run.jl

Runs!

### MITgcm_scan_output.jl

Runs!

### HS94_animation.jl

Runs!

### HS94_particles.jl

Cannot find the output of `HS94_animation.jl`

```julia
julia> include("examples/HS94_particles.jl")
ERROR: LoadError: IOError: readdir("/home/ssilvest/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/pickup_hs94.cs-32x32x5/run/"): no such file or directory (ENOENT)
Stacktrace:
 [1] uv_error
   @ ./libuv.jl:100 [inlined]
 [2] readdir(dir::String; join::Bool, sort::Bool)
   @ Base.Filesystem ./file.jl:869
 [3] readdir
   @ ./file.jl:862 [inlined]
 [4] GridLoad(γ::gcmgrid; option::String)
   @ MeshArrays ~/.julia/packages/MeshArrays/k45D3/src/Grids.jl:245
 [5] top-level scope
   @ ~/development/MITgcm.jl/examples/HS94_particles.jl:59
 [6] include(fname::String)
   @ Base.MainInclude ./client.jl:489
 [7] top-level scope
   @ REPL[8]:1
in expression starting at /home/ssilvest/development/MITgcm.jl/examples/HS94_particles.jl:56
```

The output of that scratchspace is
```julia
shell> ls /home/ssilvest/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/
MITgcm-checkpoint68y

shell> ls /home/ssilvest/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/MITgcm-checkpoint68y/
doc  eesupp  LICENSE.txt  model  pkg  README.md  tools	verification
```

### HS94_Makie.jl

same as above

## Conclusion

It looks like there is a problem with `MITgcm_run.jl` on the Macbook, but that is probably a system issue on my laptop since the example runs on the linux system.
The same for `HS94_animation.jl`. It runs on the linux system but it seems to have some pathing problem on the macbook.
For the examples dependent on `HS94_animation.jl` I could test only on the linux system and I am not sure what is happening but there seem to be an issue with the pathing.
