+++ examples/HS94_particles.jl
@@ -16,8 +16,7 @@ begin
 	writecube(fil::String,x::MeshArray) = write(fil::String,x::MeshArray)
 
 	#Path to model output
-	PICKUP_hs94_path=joinpath(MITgcmScratchSpaces.path,"pickup_hs94.cs-32x32x5")
-	pth=joinpath(PICKUP_hs94_path,"run/")
+	pth=joinpath(tempdir(),"HS94_animation_link","run")
 end
 
 # ╔═╡ fb0cacc0-4817-41cf-a4ad-ab5e5d9864f3
@@ -110,13 +109,13 @@ begin
 	    t1=m1*𝐷.dt-𝐷.dt+𝐷.t00
 	    #println(m0)
 	
-	    u0=read(𝐷.pth*𝐷.U[m0],𝐷.tmp)[:,𝐷.k]
-	    v0=read(𝐷.pth*𝐷.V[m0],𝐷.tmp)[:,𝐷.k]
+	    u0=read(joinpath(𝐷.pth,𝐷.U[m0]),𝐷.tmp)[:,𝐷.k]
+	    v0=read(joinpath(𝐷.pth,𝐷.V[m0]),𝐷.tmp)[:,𝐷.k]
 	    u0=u0.*𝐷.iDXC; v0=v0.*𝐷.iDYC; #normalize to grid units
 	    (u0,v0)=exchange(u0,v0,1) #add 1 point at each edge for u and v
 	
-	    u1=read(𝐷.pth*𝐷.U[m1],𝐷.tmp)[:,𝐷.k]
-	    v1=read(𝐷.pth*𝐷.V[m1],𝐷.tmp)[:,𝐷.k]
+	    u1=read(joinpath(𝐷.pth,𝐷.U[m1]),𝐷.tmp)[:,𝐷.k]
+	    v1=read(joinpath(𝐷.pth,𝐷.V[m1]),𝐷.tmp)[:,𝐷.k]
 	    u1=u1.*𝐷.iDXC; v1=v1.*𝐷.iDYC; #normalize to grid units
 	    (u1,v1)=exchange(u1,v1,1) #add 1 point at each edge for u and v
 	
@@ -160,7 +159,7 @@ begin
 	    plot(𝐼::Individuals)
 	Plot the initial and final positions as scatter plot in x,y plane.
 	"""
-	function plot(𝐼::Individuals)
+	function plot_xy(𝐼::Individuals)
 	    🔴_by_t = IndividualDisplacements.groupby(𝐼.🔴, :t)
 	    if (sum(names(🔴_by_t).=="lon")==0)
 	        fig=scatter(🔴_by_t[1].x,🔴_by_t[1].y,c=:red,label="t0",marker = (:circle, stroke(0)))
@@ -180,14 +179,16 @@ begin
 	∫!(𝐼)
 	
 	#Plot initial and final locations
-	f1=plot(𝐼)
+	f1=plot_xy(𝐼)
 		
 	#Carry on for several time intervals (𝐼.𝑃.𝑇 .+ 𝐷.dt and so on)
 	for tt=1:ndt-1
 	    update_FlowFields!(𝐼,𝐷)
 	    ∫!(𝐼)
 	end
-	f2=plot(𝐼)
+
+	#Plot initial and final locations (again)
+	f2=plot_xy(𝐼)
 	
 	#Add longitude and latitude
 	add_lonlat!(𝐼.🔴,𝐷.XC,𝐷.YC)
@@ -200,8 +201,8 @@ begin
 	🔴 #using this variable for sync
 	M=(timestamp=now(),author="me",run="HS94")
 	I=(metadata=M, ID=𝐼.🆔, record=🔴, position=𝐼.📌)
[…]
