### 2024-08-01 b36c099
add Darwin3 example to listing
