### 2024-08-01 369cd22
fix Makie example + improve md blocs
