+++ examples/HS94_Makie.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.43
+# v0.19.45
 
 using Markdown
 using InteractiveUtils
@@ -8,6 +8,8 @@ using InteractiveUtils
 begin
 	using JLD2, FileIO, DataFrames, CairoMakie, MITgcm
 	🏁 = "🏁"
+	pth=joinpath(tempdir(),"HS94_animation_link","run")
+	fil=joinpath(pth,"HS94.jld2") #generated by HS94_particles.jl
 end
 
 # ╔═╡ 6c682f64-d02b-4459-bfa5-9b76216dfc91
@@ -17,17 +19,17 @@ begin
 	### 
 
 
-	Here we read trajectory output generated as shown in `HS94_particles.jl` and plot them using the `Makie.jl` package to generate images and a movie.
+	Here we read trajectory output generated in `HS94_particles.jl` (file name below).
+	
+	Then we plot it using `Makie.jl` package instead of `plots.jl`.
 	"""
 end
 
 # ╔═╡ ae030b46-826a-11eb-093e-810bd32a356d
 begin
-	PICKUP_hs94_path=joinpath(MITgcmScratchSpaces.path,"pickup_hs94.cs-32x32x5")
-	fil=joinpath(PICKUP_hs94_path,"HS94.jld2") #generated by HS94_particles.jl
 	I=load(fil)["I"]
 	rec_by_t=groupby(I.record, :t)
-	🏁
+	🏁	
 end
 
 # ╔═╡ 3a2102ba-8269-11eb-14a7-af914ab4fb16
@@ -122,6 +124,11 @@ weakdeps = ["StaticArrays"]
     [deps.Adapt.extensions]
     AdaptStaticArraysExt = "StaticArrays"
 
+[[deps.AdaptivePredicates]]
+git-tree-sha1 = "7d5da5dd472490d048b081ca1bda4a7821b06456"
+uuid = "35492f91-a3bd-45ad-95db-fcad7dcfedb7"
+version = "1.1.1"
+
 [[deps.AliasTables]]
 deps = ["PtrArrays", "Random"]
 git-tree-sha1 = "9876e1e164b144ca45e9e3198d0b689cadfed9ff"
@@ -143,9 +150,9 @@ uuid = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
 
 [[deps.Automa]]
 deps = ["PrecompileTools", "TranscodingStreams"]
-git-tree-sha1 = "588e0d680ad1d7201d4c6a804dcb1cd9cba79fbb"
+git-tree-sha1 = "014bc22d6c400a7703c0f5dc1fdc302440cf88be"
 uuid = "67c07d97-cdcb-5c2c-af73-a7f9c32a568b"
-version = "1.0.3"
+version = "1.0.4"
 
 [[deps.AxisAlgorithms]]
 deps = ["LinearAlgebra", "Random", "SparseArrays", "WoodburyMatrices"]
@@ -201,9 +208,9 @@ version = "1.0.5"
 
 [[deps.CairoMakie]]
 deps = ["CRC32c", "Cairo", "Cairo_jll", "Colors", "FileIO", "FreeType", "GeometryBasics", "LinearAlgebra", "Makie", "PrecompileTools"]
-git-tree-sha1 = "3441d68ea63944a2b9b6de76603ec1c8b0fd4e3e"
+git-tree-sha1 = "e4da5095557f24713bae4c9f50e34ff4d3b959c0"
 uuid = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-version = "0.12.3"
[…]
