### 2024-08-03 gaelforget/MITgcm.jl#106 @simone-silvestri
I have tried the new version on the linux machine, and I am happy to say that all most the examples that were not running before (`MITgcm_worflow.jl` as well as the `HS94` series) seem to work.

Only two issues to solve:
- `plot_xy` in `HS94_particles.jl` is defined _after_ it is called, so I had to move it up to make sure it is defined before the first use.
- In `HS94_Makie.jl` these lines
```julia
        lo = @lift(rec_by_t[$time].lon)
        la = @lift(rec_by_t[$time].lat)

        fi1,ax1 = scatter(lo, la, markersize = 4.0, color=:blue, strokewidth = 0.0)
        xlims!(ax1,(-180.,180.)); ylims!(ax1,(-90.,90.))
```
 spit out the error:
```julia
julia> include("examples/HS94_Makie.jl")
ERROR: LoadError: Cannot convert Observable{SubArray{Float64, 1, Vector{Float64}, Tuple{Vector{Int64}}, false}} to series data for plotting
Stacktrace:
  [1] error(s::String)
    @ Base ./error.jl:35
  [2] _prepare_series_data(x::Observable{SubArray{Float64, 1, Vector{Float64}, Tuple{Vector{Int64}}, false}})
    @ RecipesPipeline ~/.julia/packages/RecipesPipeline/BGM3l/src/series.jl:8
  [3] _series_data_vector(x::Observable{SubArray{Float64, 1, Vector{Float64}, Tuple{Vector{…}}, false}}, plotattributes::Dict{Symbol, Any})
    @ RecipesPipeline ~/.julia/packages/RecipesPipeline/BGM3l/src/series.jl:36
  [4] macro expansion
    @ ~/.julia/packages/RecipesPipeline/BGM3l/src/series.jl:128 [inlined]
  [5] apply_recipe(plotattributes::AbstractDict{Symbol, Any}, ::Type{RecipesPipeline.SliceIt}, x::Any, y::Any, z::Any)
    @ RecipesPipeline ~/.julia/packages/RecipesBase/BRe07/src/RecipesBase.jl:300
  [6] _process_userrecipes!(plt::Any, plotattributes::Any, args::Any)
    @ RecipesPipeline ~/.julia/packages/RecipesPipeline/BGM3l/src/user_recipe.jl:38
  [7] recipe_pipeline!(plt::Any, plotattributes::Any, args::Any)
    @ RecipesPipeline ~/.julia/packages/RecipesPipeline/BGM3l/src/RecipesPipeline.jl:72
  [8] _plot!(plt::Plots.Plot, plotattributes::Any, args::Any)
    @ Plots ~/.julia/packages/Plots/du2dt/src/plot.jl:223
  [9] plot(::Any, ::Vararg{Any}; kw...)
    @ Plots ~/.julia/packages/Plots/du2dt/src/plot.jl:102
 [10] scatter(::Any, ::Vararg{Any}; kw...)
    @ Plots ~/.julia/packages/RecipesBase/BRe07/src/RecipesBase.jl:427
 [11] top-level scope
    @ ~/development/MITgcm.jl/examples/HS94_Makie.jl:43
 [12] include(fname::String)
    @ Base.MainInclude ./client.jl:489
 [13] top-level scope
    @ REPL[43]:1
in expression starting at /home/ssilvest/development/MITgcm.jl/examples/HS94_Makie.jl:36
Some type information was truncated. Use `show(err)` to see complete types.

```

However, these issues are not directly related with the repository and are most likely very easy to solve. So I can close this issue and mention on https://github.com/openjournals/joss-reviews/issues/6710 that my review is completed.
