### 2024-08-12 37af5b2
updates, streamline paths, etc
