+++ examples/budgets/modules.jl
@@ -79,25 +79,69 @@ read_mdsio(γ::gcmgrid,args...;kwargs...) = read(read_mdsio(args...;kwargs...),�
 
 ###
 
-diags_path="OCCA2HR1/run2"
-stat_path="dT_run_v5_stat_v5"
+#diags_path=joinpath("data","diags")
+#stats_path=joinpath("data","stats")
 
-files=glob("diags/budg*.0000000732.meta",diags_path)
-isempty(files) ? files=glob("diags/budg*.0000000024.meta",diags_path) : nothing
-isempty(files) ? error("no files found") : nothing
-
-variables=[read_meta(f).fldList for f in files]
-files=[basename(split(f,".")[1]) for f in files]
-times=glob("diags/$(files[1])*data",diags_path)
-times=parse.(Float64,[split(f,".")[end-1] for f in times])
+function files_list(diags_path="")
+ files=glob("budg*.0000000732.meta",diags_path)
+ isempty(files) ? files=glob("budg*.0000000024.meta",diags_path) : nothing
+ if isempty(files)
+  println("no budget files found")
+  variables=[]
+  files=[]
+  times=[]
+ else
+  variables=[read_meta(f).fldList for f in files]
+  files=[basename(split(f,".")[1]) for f in files]
+  times=glob("$(files[1])*data",diags_path)
+  times=parse.(Float64,[split(f,".")[end-1] for f in times])
+ end
+ files,variables,times
+ end
 
 γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
+#(files,variables,times)=files_list()
 
 ###
 
-function clim_files(year1=2004, yearN=2006)
-  lst1a=glob("diags/state_2d*data",diags_path); year1a=1980
-  lst1b=glob("*jld2",stat_path); year1b=2004
+function time_axis(ntim)
+if ntim==999
+  year_start=1941
+elseif ntim==771
+  year_start=1960
+elseif ntim==531
+  year_start=1980
+elseif ntim==240
+  year_start=1992
+else
+  year_start=0
+end
+tim=year_start .+ 1/12 .*((1:ntim).-0.5)
+
+if ntim==999
+  year_start=1941
+elseif ntim==771
+  year_start=1960
+elseif ntim==531
+  year_start=1980
+elseif ntim==240
+  year_start=1992
+else
+  year_start=0
+end
+tim=year_start .+ 1/12 .*((1:ntim).-0.5)
+
+m0=(1992-year_start)*12
+tim_clim=[(m0+m:12:m0+240) for m in 1:12]
+
+tim,tim_clim
+end
+
+###
+
+function clim_files(year1=2004, yearN=2006, diags_path="",stats_path="")
+  lst1a=glob("state_2d*data",diags_path); year1a=1980
+  lst1b=glob("*jld2",stats_path); year1b=2004
   i_a=(year1-year1a)*12 .+(1:12*(yearN-year1+1))
   i_b=(year1-year1b)*12 .+(1:12*(yearN-year1+1))
   (lst1a[i_a],lst1b[i_b])
@@ -132,28 +176,33 @@ end
 
 ###
 
-function add_one_field!(x,variable,record; fileroot="",fac=1.0)
-  if isempty(fileroot)
[…]
