+++ examples/budgets/heat_budget_main.jl
@@ -1,7 +1,22 @@
-using Pkg
-using Distributed
+
+# 0. file paths 
+
+#jul. -p 2 src/heat_budget_main.jl data/diags_budget_monthly tmp/budget_monthly_heat
+
+if !isempty(ARGS)
+  input_path=ARGS[1]
+  output_path=ARGS[2]
+else
+  input_path=joinpath("data","diags_budget_monthly")
+  output_path=joinpath("tmp","dT_budget_monthly")
+end
+
+!ispath(output_path) ? mkdir(output_path) : nothing
 
 # 1. load modules on all workers
+
+using Pkg
+using Distributed
 
 @everywhere begin
   using GLMakie, MeshArrays, JLD2
@@ -11,7 +26,10 @@ using Distributed
 
   include("modules.jl")
 
-  todo=BUDG.still_to_process()
+  input_path=$(input_path)
+  output_path=$(output_path)
+  (files,variables,times)=IO_CLIM.files_list(input_path)
+  todo=BUDG.still_to_process(output_path,times)
   println("left to do : $(length(todo))")
 
   vlist=["tend","forc","adv","dif"]
@@ -23,15 +41,16 @@ t=120
 
 ## heat budget
 
-A=BUDG.layer_heat_budget(t,1)
+A=BUDG.layer_heat_budget(t,times,input_path)
 dH=A.tend
 A_res=A.tend-A.forc-A.adv-A.dif
 
-do_plot ? fig1=PLOTS.check_budget_1(A,do_save=do_save,k=1) : nothing
+do_save ? fil1="tmp/budg_full.png" : fil1=""
+do_plot ? fig1=PLOTS.check_budget_1(A,k=1,output_file=fil1) : nothing
 
 ## mass budget
 
-C=BUDG.layer_mass_budget(t,1)
+C=BUDG.layer_mass_budget(t,times,input_path)
 dV=C.tend
 C_res=C.tend-C.forc-C.adv-C.dif
 
@@ -41,8 +60,8 @@ do_plot ? fig3=PLOTS.map_two(C.tend,C_res,cr,titles=["tend (mass)","res (mass)"]
 
 ## initial and final conditions
 
-(H,V)=BUDG.layer_snapshots(t)
-(Hp1,Vp1)=BUDG.layer_snapshots(t+1)
+(H,V)=BUDG.layer_snapshots(t,input_path)
+(Hp1,Vp1)=BUDG.layer_snapshots(t+1,input_path)
 T=H/V
 Tp1=Hp1/Vp1
 
@@ -55,7 +74,7 @@ end
 
 ## temperature tendency
 
-dt=3600*(IO_CLIM.times[t+1]-IO_CLIM.times[t])
+dt=3600*(times[t+1]-times[t])
 dTa=(dH-Tp1*dV)/V*dt
 dTb=Tp1-T
 
@@ -65,10 +84,10 @@ do_plot ? fig5=PLOTS.map_two(dTa,dTa-dTb,(-1,1).*1,
 
 # 3. main computation loop
 
-pa="tmp/dT_budget_monthly"
-!isdir(pa) ? mkdir(pa) : nothing
+!isdir(output_path) ? mkdir(output_path) : nothing
 
 @sync @distributed for t in todo
-  BUDG.temperature_equation(t,do_save=true,path=pa)
+  BUDG.temperature_equation(t,times,
+     input_path=input_path,output_path=output_path)
   GC.gc()
 end
