### 2024-09-02 #6710 @swilliamson7
> @swilliamson7 Are you near completion?

I left my comments as an issue on the MITgcm.jl repository, it hadn't been addressed yet so wasn't sure if I could say my review was complete or not. Aside from the issue I've finished going over everything (paper, code, etc.)
