### 2024-09-04 f6746e8
add docs for HS94 series, and off switches if files not found
