+++ examples/HS94_particles.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.45
+# v0.19.46
 
 using Markdown
 using InteractiveUtils
@@ -18,6 +18,14 @@ begin
 
 	#Path to model output
 	pth=joinpath(tempdir(),"HS94_animation_link","run")
+
+    found_MITgcm_files=ispath(pth)
+    if found_MITgcm_files
+        println("😄 Input files were found in \n $pth")
+    else
+        println("!!! Necessary files NOT found in \n!!! $pth")
+        println("!!! Please make sure to run the `HS94_animation.jl` notebook first")
+    end
 end
 
 # ╔═╡ 02005854-3442-4c65-910b-e8a000805d17
@@ -56,7 +64,7 @@ md"""A list of grid variables should appear afterwards if files are indeed found
 """
 
 # ╔═╡ b7c6606e-0d1a-4f65-a65a-0da320586fea
-begin
+if found_MITgcm_files
 		#Load grid files to memory
 		γ=gcmgrid(pth,"CubeSphere",6,fill((32, 32),6), [192 32], Float64, readcube, writecube)
 		Γ = GridLoad(γ,option="full")
@@ -72,8 +80,7 @@ md""" ### Setup Computation
 """		
 
 # ╔═╡ e032b6d4-959d-11eb-28df-2112db1b1e4e
-begin
-		
+if found_MITgcm_files
 	#Parameters: level, time, U/V files, grid factors
 	k=3
 	dt=7200.0
@@ -131,6 +138,8 @@ begin
 	end
 	
 	𝐼
+else
+    ndt=0
 end
 
 # ╔═╡ 5a7af9d3-b800-44fd-9ffe-0580b51ee70a
@@ -171,7 +180,7 @@ begin
 end
 
 # ╔═╡ 7cc13ee0-ad61-4274-9c94-cbfa3974046b
-begin
+if found_MITgcm_files
 	#Read velocity fields and compute trajectories
 	update_FlowFields!(𝐼,𝐷)
 	∫!(𝐼)
@@ -195,13 +204,13 @@ begin
 end
 
 # ╔═╡ a34adf36-81de-42c2-8407-d408041193b6
-f1
+found_MITgcm_files ? f1 : nothing
 
 # ╔═╡ a1fdaf62-1b80-481e-a5c7-45e82c540368
-f2
+found_MITgcm_files ? f2 : nothing
 
 # ╔═╡ c473fb90-8901-466e-a8e3-c7efd189e1ba
-f3=plot_xy(𝐼)
+found_MITgcm_files ? f3=plot_xy(𝐼) : nothing
 
 # ╔═╡ 2d320d61-9e55-471e-b663-8d95f2055b84
 md"""### Save Results
@@ -210,7 +219,7 @@ The file generated in the next code cell is used in the `HS94_Makie.jl` example.
 """
 
 # ╔═╡ d6ce6158-bdf0-48cd-8c6e-d6c1191e6a07
-begin
+if found_MITgcm_files
 	🔴 #using this variable for sync
 	M=(timestamp=now(),author="me",run="HS94")
 	I=(metadata=M, ID=𝐼.🆔, record=🔴, position=𝐼.📌)
@@ -254,9 +263,9 @@ manifest_format = "2.0"
 project_hash = "546acc5e4c6c5a847a8fb4477699191491d68a1f"
 
 [[deps.ADTypes]]
-git-tree-sha1 = "aa4d425271a914d8c4af6ad9fccb6eb3aec662c7"
+git-tree-sha1 = "99a6f5d0ce1c7c6afdb759daa30226f71c54f6b0"
[…]
