### 2024-09-04 gaelforget/MITgcm.jl#128 @gaelforget
Thanks @swilliamson7 for the very useful feedback.

> 1. The first is that HS94_Makie.jl didn't run for me. I get the error message

The reason for this is that the 3 `HS94*jl` tutorials need to be run in sequence. This is indicated in the docs but, within each notebook, the error message you get is not sufficient.

in https://github.com/gaelforget/MITgcm.jl/pull/130

I added switches to by pass running the code cells that would error out + a message that provides users direction on the notebooks they need to run first. I added another warning in the docs about this too.
