### 2024-09-04 gaelforget/MITgcm.jl#128 @gaelforget
> 2. The second, and really minor, is that in the documentation when looking through "Functionalities" a couple functions are listed as ClimateModels.function where they should be MITgcm.function, if that make sense. These are specifically the functions build, clean, compile, and setup.

I am not sure that this is an issue to fix, as these are methods that are imported from ClimateModels.jl and are specialized for MITgcm here. Also the `source` code links point to the right code within `MITgcm.jl`.

With this said, I added a statement that _The following three methods implement the [ClimateModels.jl](https://github.com/gaelforget/ClimateModels.jl/#readme)'s interface for [MITgcm](https://github.com/MITgcm/MITgcm#readme)._ just before these doc strings in the online docs.

In PR tbd ...
