### 2024-09-04 gaelforget/MITgcm.jl#128 @gaelforget
> I really like MITgcm_workflow.jl and HS94_animation.jl, since they both provide a good walkthrough of how to run specific model configurations and read the output. I do think that it would be helpful though to provide a bit more info on what's going on.

Great to hear! Thank you!!

> For example, in MITgcm_workflow.jl when we get to analyzing results the user is shown a list of (what I think are all) outputs from running the configuration, in my case this was tutorial_global_oce_biogeo. However what the variables are is sort of vague, I don't really know what's contained in something like "surfDiag.0005184004.001.002.data" and it could be useful to have a reference point on what exactly I'm looking at/what the MITgcm returns to the user. I'm guessing that this is all documented in the MITgcm documentation itself, so maybe providing a link to the relevant location?

will do

> Going to HS94_animation.jl, I have a somewhat similar comment. In the block on Workflow Steps, we're shown sort of three things. First we setup the directory, then compile, and then how to modify parameters. Setting up the directory and compiling the model is fine since this is, I'm guessing, fairly uniform across the different configurations, but the tunable parameters likely change between model configurations. It could be helpful to describe to the user what options they have, so again just a link to the MITgcm documentation and/or providing an explanation here. It’s also still not overly clear to me how one goes about changing parameters. In this example I can piece together what was done, but for example in the line
>
> fil1=joinpath(PICKUP_hs94_path,"pickup.0000043200.data”)
>
> I don’t know where the path PICKUP_hs94_path came from, or how we knew that this is how to adjust the initial state.

will do
