### 2024-09-05 34137fc
add links to MITgcm docs, plus explanation about files
