+++ examples/HS94_animation.jl
@@ -49,7 +49,8 @@ begin
 	pth_run=joinpath(myexp.folder,string(myexp.ID),"run")
 	md""" ## Model Configuration
 
-	This is a data structure of type `MITgcm_config` (a concrete type of `AbstractModelConfig`).
+	- Any MITgcm configuration can be represented as `MITgcm_config`. 
+	- This data structure is a concrete type of `AbstractModelConfig`.
 	"""
 end	
 
@@ -61,19 +62,42 @@ md""" ## Worflow Steps
 
 The animation just above results from the following workflow steps:
 
-- setup & compile
-- run model
-- process output
+- setup & build model
+- launch model run
+- look at model output
+
+Just below is the code that carries out the first two steps. Then we take a closer look at model output.  
 """
 
-# ╔═╡ 9238b863-dd69-42ce-8f36-995b4757cc1a
-md""" #### Contents of model run folder
+# ╔═╡ 0076cf3d-9fb5-426b-8cd7-5fd1d161a789
+md"""## Model Output"""
 
-The first cell below list all files found in files the run directory. The second displays the end of the standard output file (e.g. `output.txt`) generated during the model run. The third is the result of the `scan_rundir` function.
+# ╔═╡ 9238b863-dd69-42ce-8f36-995b4757cc1a
+md"""### Run folder Contents
+
+- The first cell below list all files found in files the run directory (inputs and outputs). 
+- The second displays the end of the standard output file (e.g. `output.txt`) generated during the model run. 
+"""
+
+# ╔═╡ 30c7ee99-8f99-4af0-8505-f100beca019f
+md"""!!! tip
+    The `scan_rundir` function can be used to gather summary information and key parameters like the number of model time steps. The result is shown below.
 """
 
 # ╔═╡ 88a5819f-6f35-40e8-9a82-8fd6f97001b1
-md""" #### Contents of the pickup folder"""
+md"""## Model Parameters
+
+- There are many possible combinations of run-time model parameters in MITgcm. Each optional package of MITgcm can add to the list, and require their own parameters.
+- Parameters are most often provided to MITgcm in the form of text files. However, parameters that vary in space and time are instead provided in binary or netcdf formats.
+
+!!! tip
+    For more information about specifying model parameters, please refer to the [MITgcm documentation](https://mitgcm.readthedocs.io/en/latest/getting_started/getting_started.html#parameters-simulation-controls)
+
+### Initial Conditions
+
+- The initial state of MITgcm is typically specified via `pickup` files that contain all state variables needed to restart the model from where it stopped. 
+- Here we use a file called `pickup.0000043200.data` that was previously generated. `MITgcm.jl` downloads it to `PICKUP_hs94_path`, as shown below. 
+"""
 
 # ╔═╡ 37294d8a-a70e-419a-a60b-11d09930c6b0
 readdir(PICKUP_hs94_path)
@@ -127,7 +151,7 @@ begin
 	#modify parameters to start from time step 43200, etc
 	modify_params_HS94(myexp)
 
[…]
