### 2024-09-05 gaelforget/MITgcm.jl#128 @gaelforget
> I guess in summary, providing information on the MITgcm and what capabilities it has, especially considering there will be people who come to use MITgcm.jl who (like me) have never used the MITgcm before.

agreed

> So reading pickups and building models are all sort of foreign concepts when you’re used to Julia ocean models.

I tend to think that it's not so different -- initial conditions might be in jld2, and the build step via a constructor in Julia. That's the thinking behind the design of [ClimateModels.jl](https://github.com/gaelforget/ClimateModels.jl#readme).

> As a bit of a side note, it would also be really cool if there was an example that demonstrated the adjoint capabilities the MITgcm (ECCO?) has. I’m guessing this would be computationally expensive, but still really neat to see in Julia.

Agreed, but if that's ok I'd like to separate this out as an issue for later. Adjoint examples are in fact listed in https://github.com/gaelforget/MITgcm.jl/issues/117 with other features that I feel could be good to add.
