### 2024-09-05 gaelforget/MITgcm.jl#128 @gaelforget
Maybe we are ready to close this issue? Please let me know
