### 2024-09-05 #6710 @gaelforget
> > @swilliamson7 Are you near completion?
>
> I left my comments as an issue on the MITgcm.jl repository, it hadn't been addressed yet so wasn't sure if I could say my review was complete or not. Aside from the issue I've finished going over everything (paper, code, etc.)

Sorry for the delay on my part. I have now addressed the very useful comments @ https://github.com/gaelforget/MITgcm.jl/issues/128
