### 2024-09-05 gaelforget/MITgcm.jl#128 @swilliamson7
Yes, issue can be closed! Sorry that my first comment was just me missing instructions 😅
