### 2024-09-05 gaelforget/MITgcm.jl#128 @gaelforget
> Yes, issue can be closed!

Cool!

> Sorry that my first comment was just me missing instructions 😅

Not at all. It's helped improve the notebooks & was totally fair comment. Thank you again
