+++ .github/workflows/ci.yml
@@ -19,7 +19,6 @@ jobs:
           - '1' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
         os:
           - ubuntu-latest
-          - macOS-latest
         arch:
           - x64
     steps:
