### 2024-09-19 73cbf51
proofread, update paper, add couple refs
