+++ docs/joss/paper.bib
@@ -125,11 +125,31 @@ year = {2019}
   journal = {Journal of Open Source Software}
 }
 
+@incollection{heimbach2002automatic,
+author = {Heimbach, Patrick and Hill, Chris and Giering, Ralf},
+booktitle = {Computational Science—ICCS 2002},
+pages = {1019--1028},
+publisher = {Springer},
+title = {{Automatic generation of efficient adjoint code for a parallel Navier-Stokes solver}},
+year = {2002}
+}
+
+@Article{Gallmeier2023,
+AUTHOR = {Gallmeier, K. and Prochaska, J. X. and Cornillon, P. and Menemenlis, D. and Kelm, M.},
+TITLE = {An evaluation of the LLC4320 global-ocean simulation based on the submesoscale structure of modeled sea surface temperature fields},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {16},
+YEAR = {2023},
+NUMBER = {23},
+PAGES = {7143--7170},
+URL = {https://gmd.copernicus.org/articles/16/7143/2023/},
+DOI = {10.5194/gmd-16-7143-2023}
+}
+
 @article{Forget2024,
  author = {Forget, Ga{\"{e}}l},
  title = {Energy Imbalance in the Sunlit Ocean Layer},
- journal = {Communications Earth & Environment},
- number = {submitted},
+ number = {(in prep)},
  year = {2024}
 }
 
