+++ docs/joss/paper.md
@@ -16,33 +16,33 @@ bibliography: paper.bib
 ---
 # Summary 
 
-General circulation models are used to study climate, ocean and atmosphere dynamics, biogechemistry, ecology, and more. The `MITgcm`, written in `Fortran`, is one of the most widely-used models of this class. We present `MITgcm.jl`, an interface to `MITgcm` written in Julia that can be used to not only analyze model results but also drive model simulations. With `MITgcm.jl` users can setup, build, and run `MITgcm` without having to know shell scripting or having to edit text files manually. `MITgcm.jl` provides support in `Julia` for the various input and output formats used in `MITgcm`. It implements the `ClimateModels.jl` interface, and opens up a whole new way of using `MITgcm` interactively from Jupyter and Pluto notebooks. `MITgcm.jl` in turn brings full-featured, reliable ocean modeling to Julia.
+General circulation models are used to study climate, ocean and atmosphere dynamics, biogechemistry, ecology, and more. The `MITgcm`, written in `Fortran`, is one of the most widely-used models of this class. We present `MITgcm.jl`, a two-way interface to `MITgcm` written in Julia that can be used to not only analyze model results but also drive model simulations. With `MITgcm.jl` users can setup, build, and launch `MITgcm` without having to know shell scripting or having to edit text files manually. `MITgcm.jl` provides support in `Julia` for the various input and output formats used in `MITgcm`. It implements the `ClimateModels.jl` interface, and opens up a whole new way of using `MITgcm` interactively from Jupyter and Pluto notebooks. `MITgcm.jl` in turn brings full-featured, reliable ocean modeling to Julia.
 
 # Statement of Need
 
-The cutting-edge of climate modeling, and much of its legacy, is based on numerical models written in compiled languages like `Fortran` or `C`. The MIT general circulation model (`MITgcm`) for example has recently been run globally at the kilometer scale to provide an unprecented view of ocean dynamics (Fig.~\ref{fig:examples}, @camp-etal:04,  @marshall1997fvi). With its unrivaled adjoint modeling capabilities, `MITgcm` is also the computational engine that powers the ECCO reanalysis, a data-assimilating product widely-used in climate science (Fig. \ref{fig:examples},  @Forget2015a,  @Forget2024). `MITgcm` further provides unique modeling capabilities for ocean biogeochemistry, ecology, and optics (@Dutkiewicz2015, @Cbiomes2019). While a new generation of models, written in languages like C++ and Julia, is poised to better exploit GPUs (@OceananigansJOSS, @Wu2022, @e3sm-model), `Fortran`-based models are expected to remain popular on other computer architectures for the foreseable future. They also provide a crucial reference point to evaluate next generation models.
[…]
