### 2024-09-19 #6710 @gaelforget
Probably done with edits, but planning to improve and replace the plots.
