### 2024-09-23 9a98be2
update fig1 + caption
