+++ docs/joss/paper.md
@@ -24,7 +24,7 @@ The cutting-edge of climate modeling, and much of its legacy, is based on numeri
 
 Models like `MITgcm` benefit from decades of accumulated expertise, careful inspection, and continuous integration. Running such a model or exploiting its results does not require knowing the computer language being used internally. However, `Fortran`-based models can sometime appear complicated or inconvenient to operate due to technical hurdles like having to use a compiler directly, to edit text files manually, or to deal with shell-scripting. Fortunately, such issues are easily alleviated by providing a user-friendly interface (e.g., Fig. \ref{fig:interact}) written in a high-level language (e.g., `Julia`) to interact with climate models written in lower-level languages as done here for `MITgcm`.
 
-![Examples of `MITgcm` output, read, processed, and plotted in Julia using `MITgcm.jl`. Top left : global mean ocean warming over 1980-2022 [@Forget2024]. Top right : time mean view of the global ocean conveyor belt [@Rousselet2021]. Bottom right : tracking seawater pathtways along the Gulf Stream via `IndividualDisplacements.jl` [@Forget2021]. Bottom left : global `MITgcm` simulation on a 4km resolution Lat-Lon-Cap grid [@Forget2015a] visualize in `Julia`. \label{fig:examples}](MITgcm_Examples.png){ width=100% }
+![Examples of `MITgcm` output being read, processed, plotted, and analyzed in Julia using `MITgcm.jl`. Top left : global mean ocean warming over 1980-2023 as estimated using MITgcm in OCCA2HR2 [@Forget2024]. Bottom Left : time mean view of the global ocean conveyor belt [@Rousselet2021]. Bottom right : tracking seawater pathtways along the Gulf Stream via `IndividualDisplacements.jl` [@Forget2021]. Top Right : global `MITgcm` simulation on a 4km resolution Lat-Lon-Cap grid [@Forget2015a] visualized in `Julia`. \label{fig:examples}](MITgcm_Examples.png){ width=100% }
 
 `MITgcm.jl` can read the various types of model output that `MITgcm` generates. This feature in itself enables common result analysis tasks like mapping and plotting, but also accurate computations of quantities such as ocean heat transport and global warming (Fig. \ref{fig:examples}, @Forget2015a, @Forget2019, @Forget2024). In addition, `MITgcm.jl` makes it easy to deploy and run any configuration of `MITgcm` on laptops, HPC clusters, and in the cloud. `MITgcm.jl` interacts with `MITgcm`'s run-time model parameters. These are represented in `Julia` as an ordered dictionary, and can be re-exported to the standard `TOML` file format, or to the native `MITgcm` format. Owing to this two-way interface, `MITgcm` can now be used from Jupyter or Pluto notebooks directly via `MITgcm.jl`. The package includes a series of examples and tutorials that demonstrate the interface (e.g., Fig. \ref{fig:interact}). 
 
