### 2024-09-23 #6710 @gaelforget
> @gaelforget: do you need additional time to improve and replace the plots? If so, could you give me a rough time estimate of when you will be done?

All set with plots. Thank you for your patience @AnjaliSandip
