### 2024-09-23 84593df
improve fig1 caption, add a bit of detail
