+++ docs/joss/paper.md
@@ -26,9 +26,9 @@ Models like `MITgcm` benefit from decades of accumulated expertise, careful insp
 
 ![Examples of `MITgcm` output created, read, processed, plotted, and analyzed via `MITgcm.jl` in `Julia`. Top left : global mean ocean warming over 1980-2023 as simulated in `OCCA2HR2` using `MITgcm.jl` [@Forget2024]. Bottom Left : climatological mean view of the global ocean conveyor belt [@Rousselet2021] produced in `Climatology.jl`. Bottom right : particle tracking of seawater pathtways along the Gulf Stream, computed in three dimensions by `IndividualDisplacements.jl` [@Forget2021]. Top Right : temperature fronts in a global `MITgcm` simulation, which ran on a 4km resolution Lat-Lon-Cap grid [@Forget2015a,Gallmeier2023], diagnozed and visualized in `Julia` using `MeshArrays.jl`. \label{fig:examples}](MITgcm_Examples.png){ width=100% }
 
-`MITgcm.jl` can read the various types of model output that `MITgcm` generates. This feature in itself enables common result analysis tasks like mapping and plotting, but also accurate computations of quantities such as ocean heat transport and global warming (Fig. \ref{fig:examples}, @Forget2015a, @Forget2019, @Forget2024). In addition, `MITgcm.jl` makes it easy to deploy and run any configuration of `MITgcm` on laptops, HPC clusters, and in the cloud. `MITgcm.jl` interacts with `MITgcm`'s run-time model parameters. These are represented in `Julia` as an ordered dictionary, and can be re-exported to the standard `TOML` file format, or to the native `MITgcm` format. Owing to this two-way interface, `MITgcm` can now be used from Jupyter or Pluto notebooks directly via `MITgcm.jl`. The package includes a series of examples and tutorials that demonstrate the interface (e.g., Fig. \ref{fig:interact}). 
+`MITgcm.jl` can read the various types of output that `MITgcm` generates. This feature enables not only common analyses of model results (e.g., mapping and plotting), but also accurate computations of quantities such as ocean heat transport and global warming (Fig. \ref{fig:examples}, @Forget2015a, @Forget2019, @Forget2024). `MITgcm.jl` also makes it easy to deploy and run any configuration of `MITgcm` on laptops, HPC clusters, and in the cloud. `MITgcm.jl` interacts with `MITgcm`'s run-time model parameters in `Julia` via ordered dictionaries. It can re-export them to the standard `TOML` file format, or to the native `MITgcm` format. Owing to this two-way interface, `MITgcm` can now be used from Jupyter or Pluto notebooks interactively via `MITgcm.jl`. The package includes a series of examples and tutorials that demonstrate the interface (e.g., Fig. \ref{fig:interact}). 
 
-In the code example below, `MITgcm_config` defines the `MC` data structure. The `run` command is equivalent to the `setup(MC); build(MC); launch(MC)` sequence. And `readdir` is used to list the content of the folder where the model ran (a temporary folder by default).
+In the code example below, `MITgcm_config` defines the `MC` data structure. The `run` command is equivalent to the sequence of `setup(MC); build(MC); launch(MC)`. And `readdir` returns a list of files found in the folder where `MITgcm` ran (a temporary folder by default).
 
 ```
 using MITgcm
