### 2024-09-24 #6710 @gaelforget
Did my final round of proofreading of text and plots. Good to go 😄
