### 2024-09-25 #6710 @gaelforget
> @simone-silvestri @swilliamson7 Could you please review the changes made to the paper and let me know if it is ready for publication?

In case this is helpful, I created this diff that shows the changes made to the paper (files in the joss folder) since version `0.4.2` (Jun 23):

https://github.com/gaelforget/MITgcm.jl/compare/7f9877117a00ba67921324999c24ae7a56a9084e..581e1abc3e1b7ec1a9246ec84cb20c55b08eef7f
