### 2024-09-27 #6710 @swilliamson7
> @simone-silvestri @swilliamson7 Could you please review the changes made to the paper and let me know if it is ready for publication?

@AnjaliSandip I think it's ready for publication

And @gaelforget Thank you for creating the diff file! Definitely helped
