### 2024-10-07 #6710 @AnjaliSandip
@simone-silvestri  Could you please review the changes made to the paper and let me know if it is ready for publication?
