### 2024-10-07 #6710 @AnjaliSandip
@gaelforget  Can you make a new release (if possible) of the software being reviewed and deposit a new archive of the software with Zenodo/figshare? After you have, please update the review thread with this archive DOI: `@editorialbot set 10.5281/zenodo.xxxxxx as archive`.
