+++ test/runtests.jl
@@ -16,7 +16,7 @@ using MITgcm.ClimateModels.CSV
     @test isfile(fil)
 
     #format conversions
-    (γ,Γ)=MeshArrays.GridOfOnes("CubeSphere",30,30)
+    (γ,Γ)=Grids_simple.GridOfOnes("CubeSphere",30,30)
     @test isa(convert2gcmfaces(Γ.XC),Array)
     @test isa(convert2array(Γ.XC),Array)
 
