### 2024-10-07 291a5b0
eliminate Climatology dependency in test (Climatology depends on MITgcm)
