### 2024-10-09 d588518
update notebook depedencies
