+++ examples/HS94_particles.jl
@@ -148,37 +148,6 @@ md""" ### Trajectory Computation
 The particle trajectory output is in tabular format. 
 """
 
-# ╔═╡ 5129ac14-0861-40b5-b97e-0d0edd37849f
-md""" ### Plot Results
-
-The plots below, generated by `plot_xy` function, show:
-
-- 1. initial positions (red) and after 1 day (blue; `x,y`)
-- 2. initial and final positions ($ndt days; `x,y`)
-- 3. initial and final coordinates ($ndt days; `lon,lat`)
-
-To save a plot, add somthing like `savefig(tempdir()*"/"*"hs94.cs.particles.png")`
-"""
-
-# ╔═╡ e9535b96-07c2-4aff-964d-bb327c5f735b
-begin
-	"""
-	    plot(𝐼::Individuals)
-	Plot the initial and final positions as scatter plot in x,y plane.
-	"""
-	function plot_xy(𝐼::Individuals)
-	    🔴_by_t = IndividualDisplacements.groupby(𝐼.🔴, :t)
-	    if (sum(names(🔴_by_t).=="lon")==0)
-	        fig=scatter(🔴_by_t[1].x,🔴_by_t[1].y,c=:red,label="t0",marker = (:circle, stroke(0)))
-	        scatter!(🔴_by_t[end].x,🔴_by_t[end].y,c=:blue,label="t1",marker = (:circle, stroke(0)))
-	    else
-	        fig=scatter(🔴_by_t[1].lon,🔴_by_t[1].lat,c=:red,label="t0",marker = (:circle, stroke(0)))
-	        scatter!(🔴_by_t[end].lon,🔴_by_t[end].lat,c=:blue,label="t1",marker = (:circle, stroke(0)))
-	    end
-	    return fig
-	end
-end
-
 # ╔═╡ 7cc13ee0-ad61-4274-9c94-cbfa3974046b
 if found_MITgcm_files
 	#Read velocity fields and compute trajectories
@@ -203,6 +172,18 @@ if found_MITgcm_files
 	🔴 = 𝐼.🔴
 end
 
+# ╔═╡ 5129ac14-0861-40b5-b97e-0d0edd37849f
+md""" ### Plot Results
+
+The plots below, generated by `plot_xy` function, show:
+
+- 1. initial positions (red) and after 1 day (blue; `x,y`)
+- 2. initial and final positions ($ndt days; `x,y`)
+- 3. initial and final coordinates ($ndt days; `lon,lat`)
+
+To save a plot, add somthing like `savefig(tempdir()*"/"*"hs94.cs.particles.png")`
+"""
+
 # ╔═╡ a34adf36-81de-42c2-8407-d408041193b6
 found_MITgcm_files ? f1 : nothing
 
@@ -211,6 +192,25 @@ found_MITgcm_files ? f2 : nothing
 
 # ╔═╡ c473fb90-8901-466e-a8e3-c7efd189e1ba
 found_MITgcm_files ? f3=plot_xy(𝐼) : nothing
+
+# ╔═╡ e9535b96-07c2-4aff-964d-bb327c5f735b
+begin
+	"""
+	    plot(𝐼::Individuals)
+	Plot the initial and final positions as scatter plot in x,y plane.
+	"""
+	function plot_xy(𝐼::Individuals)
+	    🔴_by_t = IndividualDisplacements.groupby(𝐼.🔴, :t)
+	    if (sum(names(🔴_by_t).=="lon")==0)
[…]
