### 2024-10-11 #6710 @AnjaliSandip
@gaelforget Can you fix the missing DOIs?
