+++ docs/joss/paper.bib
@@ -131,6 +131,7 @@ booktitle = {Computational Science—ICCS 2002},
 pages = {1019--1028},
 publisher = {Springer},
 title = {{Automatic generation of efficient adjoint code for a parallel Navier-Stokes solver}},
+doi = {https://doi.org/10.1007/3-540-46080-2_107},
 year = {2002}
 }
 
@@ -150,6 +151,7 @@ DOI = {10.5194/gmd-16-7143-2023}
  author = {Forget, Ga{\"{e}}l},
  title = {Energy Imbalance in the Sunlit Ocean Layer},
  number = {(in prep)},
+ doi = {https://doi.org/10.21203/rs.3.rs-3979671/v1},
  year = {2024}
 }
 
@@ -174,6 +176,7 @@ number = {C3},
 pages = {5753--5766},
 title = {{A finite-volume, incompressible Navier Stokes model for studies of the ocean on parallel computers}},
 volume = {102},
+doi = {https://doi.org/10.1029/96JC02775},
 year = {1997}
 }
 
@@ -183,18 +186,7 @@ journal = {Ocean Modelling},
 pages = {221--244},
 title = {{Conservation of properties in a free surface model}},
 volume = {6},
+doi = {https://doi.org/10.1016/S1463-5003(03)00009-X},
 year = {2004}
 }
-
-@article{HHG05,
-author = {Heimbach, P and Hill, C and Giering, R},
-journal = {Future Generation Computer Systems},
-number = {8},
-pages = {1356--1371},
-publisher = {Elsevier},
-title = {{An efficient exact adjoint of the parallel MIT general circulation model, generated via automatic differentiation}},
-volume = {21},
-year = {2005}
-}
-
 
