### 2024-10-17 #6710 @gaelforget
> @gaelforget Can you fix the missing DOIs?

Should be all set following https://github.com/gaelforget/MITgcm.jl/pull/139

However I am not able to recompile paper via the bot. Works on my laptop but not here for some reason that I do not understand.
