### 2024-10-17 #6710 @gaelforget
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/11386076261?check_suite_focus=true).

Is this an issue on your end or could you explain how to fix on my end?

The DOIs should be all set as mentioned in https://github.com/openjournals/joss-reviews/issues/6710#issuecomment-2418521418
