### 2024-10-18 #6710 @tarleb
@AnjaliSandip please rerun

````
```

The system got confused by the link, the DOI should be just a plain string.
