### 2024-10-21 #6710 @kyleniemeyer
@gaelforget the previous issue has been resolved, but now it looks like there is a minor issue in the bibliography database with duplicate cite keys- can you address that?
