+++ docs/joss/paper.bib
@@ -110,21 +110,6 @@ year = {2019}
   journal = {Journal of Open Source Software}
 }
 
-@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Ga{\"{e}}l Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
-
-@article{Forget2021,
-  doi = {10.21105/joss.02813},
-  url = {https://doi.org/10.21105/joss.02813},
-  year = {2021},
-  publisher = {The Open Journal},
-  volume = {6},
-  number = {60},
-  pages = {2813},
-  author = {Ga{\"{e}}l Forget},
-  title = {IndividualDisplacements.jl: a Julia package to simulate and study particle displacements within the climate system},
-  journal = {Journal of Open Source Software}
-}
-
 @incollection{heimbach2002automatic,
 author = {Heimbach, Patrick and Hill, Chris and Giering, Ralf},
 booktitle = {Computational Science—ICCS 2002},
