### 2024-10-21 #6710 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
