### 2024-10-21 #6710 @gaelforget
> If the paper PDF and the deposit XML files look good ...

PDF looks good to go. Have not checked the XML
