### 2024-10-22 #6710 @kthyng
@gaelforget Can you update the title and author list in the Zenodo archive to exactly match the JOSS paper?
