### 2024-10-22 #6710 @kthyng
Paper edits: https://github.com/gaelforget/MITgcm.jl/pull/141
