### 2024-10-22 #6710 @gaelforget
> Paper edits: [gaelforget/MITgcm.jl#141](https://github.com/gaelforget/MITgcm.jl/pull/141)

merged
