### 2024-10-22 #6710 @gaelforget
> @gaelforget Can you update the title and author list in the Zenodo archive to exactly match the JOSS paper?

done (see https://zenodo.org/records/13900906)
