### 2024-10-22 #6710 @kthyng
@gaelforget If you'd like to join JOSS as a reviewer, please sign up here: https://reviewers.joss.theoj.org/!
